### 2019-09-30 #344 @karinorman
submission: taxadb
Submitting Author: Kari Norman (@karinorman)
Repository:   https://github.com/cboettig/taxadb
Version submitted:   v1.0.0
Editor: @ldecicco-USGS (Guest Editor)
Version accepted: TBD

---

```
Package: taxadb
Version: 0.0.1
Title: A High-Performance Local Taxonomic Database Interface
Description: Creates a local database of many commonly used taxonomic authorities
             and provides functions that can quickly query this data.
            person("Carl", "Boettiger", , "cboettig@gmail.com", c("aut", "cre"),
                   comment = c(ORCID = "0000-0002-1642-628X")),
            person("Kari", "Norman", role= "aut",
            person("Jorrit", "Poelen", role ="aut",
                   comment = c(ORCID="0000-0003-3138-4118")),
            person("Scott", "Chamberlain", role ="aut",
                   comment = c(ORCID="0000-0003-1444-9135")),
            person("Noam", "Ross", , "ross@ecohealthalliance.org", role ="ctb",
            )
    DBI,
    arkdb,
    memoise,
    rappdirs,
    stringi,
    progress,
    dbplyr
    covr,
    RSQLite,
    MonetDBLite
Remotes: MonetDB/MonetDBLite-R, cwida/duckdb/tools/rpkg

```

The package accesses existing data sets, adapts them to a standard format, and provides a set of functions for accessing the resulting database.

Anyone using data including species names, especially for datasets that may have taxonomic inconsistencies.

`taxaize` provides accesses to many of the same datasets via an API interface. Our package is specifically designed to work around the API to offer the faster access necessary for large datasets, as well as provide a standardized format for all datasets.
