### 2019-10-03 ee2b5ee
Add citations, render them in table.
