+++ paper/refs.bib
@@ -158,16 +158,52 @@ and Logs},
 	pages = {4701--4735}
 }
 
-@article{Wieczorek2012, 
-  title={Darwin Core: An Evolving Community-Developed Biodiversity Data Standard}, 
-  volume={7}, 
+@article{Wieczorek2012,
+  title={Darwin Core: An Evolving Community-Developed Biodiversity Data Standard},
+  volume={7},
   url = {https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0029715},
-  DOI={10.1371/journal.pone.0029715}, 
+  DOI={10.1371/journal.pone.0029715},
   language = {en},
-  number={1}, 
-  journal={PLoS ONE}, 
-  author={Wieczorek, John and Bloom, David and Guralnick, Robert and Blum, Stan and Döring, Markus and Giovanni, Renato and Robertson, Tim and Vieglais, David}, 
-  year={2012}, 
+  number={1},
+  journal={PLoS ONE},
+  author={Wieczorek, John and Bloom, David and Guralnick, Robert and Blum, Stan and Döring, Markus and Giovanni, Renato and Robertson, Tim and Vieglais, David},
+  year={2012},
   month={Jun}
 }
-  
+
+@article{Rees2017,
+	author = {Jonathan A. Rees and Karen Cranston},
+	title = {Automated assembly of a reference taxonomy for phylogenetic data synthesis},
+	volume = {5},
+	number = {},
+	year = {2017},
+	doi = {10.3897/BDJ.5.e12581},
+	publisher = {Pensoft Publishers},
+	abstract = {Taxonomy and nomenclature data are critical for any project that synthesizes biodiversity data, as most biodiversity data sets use taxonomic names to identify taxa. Open Tree of Life is one such project, synthesizing sets of published phylogenetic trees into comprehensive summary trees. No single published taxonomy met the taxonomic and nomenclatural needs of the project. Here we describe a system for reproducibly combining several source taxonomies into a synthetic taxonomy, and we discuss the challenges of taxonomic and nomenclatural synthesis for downstream biodiversity projects.},
+	issn = {1314-2836},
+	pages = {e12581},
+	URL = {https://doi.org/10.3897/BDJ.5.e12581},
+	eprint = {https://doi.org/10.3897/BDJ.5.e12581},
+	journal = {Biodiversity Data Journal}
+}
+
+@misc{gbif,
+	title = {GBIF.org},
+	url = {https://www.gbif.org},
+	author = {GBIF},
+	year = {2019}
+}
+
+@misc{fishbase,
+	title = {FishBase},
+	url = {https://www.fishbase.org},
+	author = {Froese, R. and Pauly, D},
+	year = {2019}
+}
+
+@misc{wikidata,
+	title = {GBIF.org},
+	url = {https://www.gbif.org},
+	author = {GBIF},
+	year = {2019}
+}
