### 2019-10-03 #344 @melvidoni
Hello @karinorman, your editor will be @ldecicco-USGS, acting as a Guest Editor for rOpenSci.
