### 2019-10-07 #344 @ldecicco-USGS
### Editor checks:

---

#### Editor comments

Hello @karinorman ...this is my first "guest edit"...so bear with me! The following is the output of goodpractice::gp(). I'm not terribly concerned about the messages, but they would be easy messages to get rid of, and then future pull request could use the `gp()` function from a clean slate.

```
── GP taxadb ──────────────────────────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all
    package code in general. 83% of code lines are covered
    by test cases.

    R/clean_names.R:42:NA
    R/clean_names.R:43:NA
    R/clean_names.R:44:NA
    R/clean_names.R:45:NA
    R/clean_names.R:46:NA
    ... and 60 more lines

  ✖ use '<-' for assignment instead of '='. '<-'
    is the standard, and R users and developers are used it
    and it is easier to read your code for them if you use
    '<-'.

    R\handling-duplicates.R:39:11
    R\mutate_db.R:56:24
    R\mutate_db.R:69:25

  ✖ avoid long code lines, it is bad for
    readability. Also, many people prefer editor windows
    that are about 80 characters wide. Try make your lines
    shorter than 80 characters

    inst\examples\docker.R:23:1
    R\by_common.R:4:1
    R\clean_names.R:3:1
    R\clean_names.R:20:1
    R\filter_by.R:12:1
    ... and 24 more lines

  ✖ avoid 1:length(...), 1:nrow(...),
    1:ncol(...), 1:NROW(...) and 1:NCOL(...) expressions.
    They are error prone and result 1:0 if the expression
    on the right hand side is zero. Use seq_len() or
    seq_along() instead.

    R\filter_by.R:60:50
    R\filter_by.R:65:49
    tests\testthat\test-mutate-db.R:16:19

──────────────────────────────────────────────────────────────
```
---
I'm sending email requests now to find 2 reviewers.

Reviewers:
Due date:
