### 2019-10-11 50a9c0b
Add descriptions to the provider table.
