### 2019-10-11 bbe4140
Address gp() warnings, mostly linebreaks but also some new tests.
