### 2019-10-11 061a933
Address gp() warnings, mostly linebreaks but also some new tests.
