+++ R/by_common.R
@@ -1,7 +1,8 @@
 
 #' Look up taxonomic information by common name
 #'
-#' @param name a character vector of common (vernacular English) names, e.g. "Humans"
+#' @param name a character vector of common (vernacular English) names,
+#' e.g. "Humans"
 #' @inheritParams filter_by
 #' @return a data.frame in the Darwin Core tabular format containing the
 #' matching taxonomic entities.
@@ -41,7 +42,7 @@ by_common <- function(name,
 NO_COMMON <- c("common_wd", "common_ott", "common_tpl")
 
 assert_has_common <- function(provider){
-  
+
   if (paste0("common_", provider) %in% NO_COMMON){
     warning(paste("taxadb provider", provider,
                          "does not provide common names at this time."),
