+++ tests/testthat/test-clean_names.R
@@ -1,5 +1,15 @@
 context("clean_names")
 
+test_that("set_space_delim", {
+  x <- set_space_delim("Poa annua. ")
+  expect_identical(x, "Poa annua")
+})
+
+test_that("drop_sp.", {
+  x <- drop_sp.("Poa sp.")
+  expect_identical(x, "Poa")
+})
+
 test_that("binomial_names", {
   x <- binomial_names("Homo")
   expect_identical(x, "Homo")
@@ -14,3 +24,7 @@ test_that("drop_parenthetical", {
   drop_parenthetical("Poa annua (Smith 1912)")
 })
 
+test_that("clean_names", {
+  x <- clean_names(" Poa annua sp. (Smith 1912)", lowercase = FALSE)
+  expect_identical(x, "Poa annua")
+})
