+++ man/name_starts_with.Rd
@@ -7,7 +7,8 @@
 name_starts_with(name, provider, db = td_connect(), ignore_case = TRUE)
 }
 \arguments{
-\item{name}{vector of names (scientific or common, see \code{by}) to be matched against.}
+\item{name}{vector of names (scientific or common, see \code{by})
+to be matched against.}
 
 \item{provider}{from which provider should the hierarchy be returned?
 Default is 'itis'.}
