### 2019-10-28 #344 @lindsayplatt
Hello! It was fun to review this package. I have specific comments within the ROpenSci checklist (they are noted as a quotation starting with my initials `{LP}`). I also added a section of comments near the end that didn't fit the checklist perfectly.

Best,
Lindsay

> {LP} I see a vignettes folder on GitHub, but I don't see the vignettes when I go to the Packages tab and click on the package name or when I run `vignette(package="taxadb")`, I get `no vignettes found`. There seems to be an issue with finding the vignettes that are in the folder. It might be that you have them tucked under an additional folder (`articles`) rather than the .Rmd files just existing in the `vignettes` folder.

- [ ] **Examples** for all exported functions in R Help that run successfully locally.

> {LP} `mutate_db()` & `taxa_tbl()` do not have examples. `td_connect()` and `td_create()` have examples but using functions called `db_connect()` and `db_create()`; perhaps outdated function names? All others have examples and they run locally.

> {LP} I do not see anything that mentions how to contribute to the package in either the README, in a file called CONTRIBUTING, or mentioned anywhere in the package files. Please see this part of the ROpenSci package documentation (https://devguide.ropensci.org/collaboration.html#friendlyfiles) for info on including a contribution section. The DESCRIPTION file has the required fields.

_**Have not heard if this is being submitted to JOSS, so assuming no and skipping the JOSS section of review.**_

> {LP} Yes, connects to local DB of taxonomic data for querying consistently. Unknown information about the source of this data, which is useful in case you need to update frequently by running `td_create("col")`. You should site the "Catalogue of Life" somewhere, include links to documentation for that, and then tell the user how often (if at all) they should update that database by running `td_create`.

> {LP} Since the database is available locally, it seems like this package could be faster than it currently is. Not bad right now though.

> {LP} I used `devtools::check()` - it completed with zero issues (I have Windows 10) and took 7.5 min. I also ran `devtools::test()` and that took 10 min with 0 failures, 5 warnings, and 2 skips. With `covr::package_coverage()`, I'm seeing low coverage for code within `mutate_db.R` and I also suggest that you add tests for `quick_db()` and `has_tbl` on their own. I ran `goodpractice::gp()` and only got one note - it was about test coverage (87% lines covered) with similar information about some missing lines as I noted previously. Finally, I ran `spelling::spell_check_package()` and `spelling::spell_check_files("README.Rmd")` with no real spelling errors.

> {LP} Two things, both of which have been noted above. All else check out! One issue - vignettes are not available to the user since they aren't in the correct folder. One suggestion - watch the naming convention. Your `by_*` functions are a bit confusing and could be made clearer with more descriptive names.

Estimated hours spent reviewing: 2.5 hrs as of 10/27/2019

---

#### README Documentation

* In the installation instructions, it's unclear which of these steps is required as part of the setup and which is an example. It seems like if the package needs `dplyr` and `readr` that they would be loaded when `taxadb` is loaded, rather than requiring additional `library` commands. You might want to explain why those are needed. Additionally, the local copy of the Catalogue of Life seems like a necessary step, but the following code about reading in a species list seems like an example and not required for all usecases. I would just clarify when/why the user might take those steps.
* Using "col" as a reference to "Catalogue of Life" is potentially confusing. I like the brevity of it, but it has a lot of other uses in R such as "column" in `colnames` or specifying "color" in base graphics. Consider using something more descriptive, such as "cat_of_life".
* For people that may be new to the Catalogue of Life, it would be useful to have some documentation describing the common column names or pointing people out to existing documentation for this database.
* When I installed this, I received some warnings. Better warnings/errors to help me understand if I need to take action or try again would be useful.

```r
td_create("col")

Importing dwc_col.tsv.bz2 in 100000 line chunks:
[-] chunk 6	...Done! (in 47.58484 secs)
Importing common_col.tsv.bz2 in 100000 line chunks:
Warning: 1 parsing failure.
  row col   expected    actual         file
23215  -- 20 columns 2 columns literal data

	...Done! (in 2.237027 secs)
1: In readLines(con, n = n, encoding = encoding, warn = FALSE) :
  invalid input found on input connection 'C:\Users\lcarr\AppData\Local\taxadb\taxadb/dwc_col.tsv.bz2'
2: In readLines(con, n = n, encoding = encoding, warn = FALSE) :
  invalid input found on input connection 'C:\Users\lcarr\AppData\Local\taxadb\taxadb/common_col.tsv.bz2'
```

#### Functions

* When I ran the `by_*` functions, most of the examples took time to download information. If you submit to CRAN, these would not pass checks because they take too long for the system to run. I recommend that you use `\dontrun{}` in your ROXYGEN documentation for the examples that take a minute to download or hit some kind of website.
* I would suggest having the `by_*` family of functions be more descriptive. Typically, functions are named with a verb indicating what they do. For instance, `filter_catalogue_by*`. It is longer, but brevity can cause more confusion because it is not as clear what action the function is actually taking. It could be "grouping by" or "summarizing by", etc.
* It is unclear to me how `filter_by` is different from `by_*`. I can run the following two lines and seem to get the same result:

```r
by_name(c("Trochalopteron henrici gucenense", "Trochalopteron elliotii"))

filter_by(c("Trochalopteron henrici gucenense", "Trochalopteron elliotii"), "scientificName")
```

* It could be a little confusing that `get_ids` has an argument that can completely change the output format (the `format = "uri"` argument). Typically, no matter the arguments, you can expect the same type of information back. Returning a URL that you need to visit to get the IDs might be surprising to users. You could make it a separate function. Additionally, if I try that same argument for `get_names`, I don't get a url back - `get_names(c("ITIS:180092", "ITIS:179913"), format = "uri")` returns `"Homo sapiens" "Mammalia"`.
* I see code for `null_tibble()` but I don't see it being used at all or exported. Might be old and you don't need it anymore?

#### Miscellaneous

* If I run code in the console and then try to knit an RMD with `taxadb` commands, I always get the error about the database being locked since there is already a session using it. It isn't super clear to a user how to remedy this without completely closing R, opening it up, and trying to knit the RMD without running any code in the console. It may make sense to include a function that the user can run to "close" the connection manually and also explain this behavior in documentation somewhere.
* I don't know how to recreate or describe (and it has not happened to me before), but every once in a while as I am executing commands from this package, RStudio seems to freeze. I can't click on anything or type any code. I have to close the window and reopen to keep going. I seem to be able to close the window just fine; no need to force quit, so it isn't totally frozen. Haven't seen this happen to me before. I am using RStudio version 1.2.5001
