### 2019-10-28 #344 @ldecicco-USGS
Thanks @lindsayplatt !

It looks like the vignettes are in a sub-directory "articles", which is why they are not being properly indexed when created via the `remotes` package.

If the developers want the vignettes included in the package itself (and not just a pkgdown page), they would need to both move the files out of the "articles" folder, and change the installation instructions to:

```r
remotes::install_github("cboettig/taxadb", build_vignettes = TRUE, build_opts = c("--no-resave-data", "--no-manual"))
```

I'm constantly having to remind myself of the build_opts argument.
