### 2019-10-29 #344 @mcsiple
This package is very cool. I tried to follow similar conventions in my review to @lindsayplatt, for consistency. I've also added some comments on the end. Thank you for including me as a reviewer for this nice new tool.

> One quick question about the README -- I wasn't sure after reading it whether there is a way for users to know where the local database is located. If this option is available, I think it would be good to add.
>	`browseVignettes(package="taxadb")` does not return any vignettes but maybe this is normal for packages not yet on CRAN? I see that @lindsayplatt ran into a similar issue and looks like it will be resolved with @ldecicco-USGS 's  suggestion.
> I don't see any CONTRIBUTING guidelines in the documentation.

>#### Functionality

>##### Comment on operating systems
>I tested installation from GitHub and locally from two different operating systems. Where possible, I tested other functionality was also tested on both a Mac and a Windows OS.

>I ran into some issues with functionality when using different machines. For example, I had namespace errors for `rlang` and other dependencies when testing on a Windows OS. This may be unrelated to the package itself (maybe has to do with package versions?) but just something to be aware of.

>###### Local install
>Local install in Windows got stuck in a loop of error messages every time I tried it, and I wasn't able to figure out what was causing the issue, but all the error messages were related to MonetDBLite. Installation of MonetDBLite from GitHub was smooth (though lengthy) on a Mac machine.
>Local installation on a Mac ran smoothly. I'm not sure if that's an issue of different versions of the imported packages on those two machines, or something else.
> Prior to running `test()`, I installed the dev version of MonetDBLite from GitHub. The most current version did not install from CRAN (i.e., `install.packages("MonetDBLite")` returned an error). It is possible that if folks try to download and run taxadb, they may have to download the dev version of MonetDBLite in order to install taxadb. That was my experience.

>###### Install from GitHub
> `devtools::install_github("cboettig/taxadb")` was fine and fast on both Windows and Mac OS.

> #### Functionality notes
> `td_create()` returned some errors when I used one machine, including some parsing failures. After the parsing failures I switched to just testing on Mac OS, and this step works fine. I wrote more below about specific functions.

> Initially `test()` returned errors because of the issues with MonetDBLite. After installing the dev version of MonetDBLite it works (and quickly!) in both operating systems.
> I used `goodpractice::gp()` too and encountered two `goodpractice` errors:

1.  Writing unit tests for all functions: "87% of code lines are covered by test cases"

2. Installation fail for R CMD check error. gp output says "Installation failed. See `~/taxadb.Rcheck/00install.out` for details"

> I never figured out what the latter issue was, but it did not appear to affect functionality for me.
> There must be some coverage issues, and `covr::package_coverage()` returns a more detailed description:
```r
taxadb Coverage: 87.09%
R/mutate_db.R: 0.00%
R/taxa_tbl.R: 79.17%
R/td_connect.R: 86.96%
R/get_ids.R: 87.50%
R/get_names.R: 90.00%
R/fuzzy_filter.R: 94.00%
R/td_create.R: 94.74%
R/clean_names.R: 95.45%
R/synonyms.R: 97.06%
R/by_common.R: 100.00%
R/by_id.R: 100.00%
R/by_name.R: 100.00%
R/by_rank.R: 100.00%
R/filter_by.R: 100.00%
R/handling-duplicates.R: 100.00%
R/utils.R: 100.00%
```

Estimated hours spent reviewing: 10 hours as of 10/28/2019

---

As someone who works mostly with one group of species and needs access to multiple databases, this package could be incredibly helpful. I think some changes to the documentation could make it more accessible for someone like myself, who is not very well-versed in the diversity of taxonomic databases, how to access them, and which ones have what I need.

The documentation is generally good and I just had a couple of small comments.

* The breeding bird survey data is used for all the intro examples, but I would love a couple of other examples that are more general. For example, looking up another set of taxa, or a set of taxa downstream from a specific taxonomic node.
* An example where taxadb interacts with another package would be helpful
* I think some documentation could be particularly effective if it is built in as more specific error/warning messages. With all the taxonomy packages I've used, I often find it hard to match exactly what the package expects in terms of the spelling and syntax of species names, especially common names. It would be great to have a little more built-in functionality that provides additional guideance (e.g., "this species is in ITIS but not in IUCN - did you mean `"itis"`?")

### Help files
Running `help(package = "taxadb")` did not return a package index or any help files when I was using Windows; on Mac OS it was fine.

## Functions
I think I may have missed something about how taxadb handles common names. I am always looking for taxonomic packages that can deal with slight differences in how common names are spelled/punctuated. It would be good to have some more flexibility w/r/t names, so that punctuation can be not an "exact match." For example:

```r
by_common("Stellers jay")
by_common("Steller jay")
by_common("Steller's jay") # only this one returns results
# Based on what I understand from clean_names, maybe this works?
by_common(clean_names(names = "Steller jay"))  # but this does not return results either
```
Maybe there is a way to set tolerance on punctuation-- if so, it would be useful for me as a future user.

In the by_rank() function, is there a way to return suggestions about what went wrong, if `by_rank()` doesn't return a tibble? For example, if I, a fish newb, am looking for all Chondrichtyans,
```r
by_rank("Osteichthyes", "class", "itis")
```
returns an empty tibble. It would be great to have something that tells people, "[insert name of taxon you entered] is defined as an [order/kingdom/genus]-- is that the classification you meant?" It doesn't have to be that literal, but would be good as a checker if people have errors in classifying the taxa they're interested in.

## Session Info
In Windows:
```r
Platform: x86_64-w64-mingw32/x64 (64-bit)
Running under: Windows >= 8 x64 (build 9200)

[1] LC_COLLATE=English_United States.1252  LC_CTYPE=English_United States.1252
[3] LC_MONETARY=English_United States.1252 LC_NUMERIC=C
[5] LC_TIME=English_United States.1252

[1] readr_1.3.1  dplyr_0.8.3  taxadb_0.0.1

 [1] Rcpp_1.0.2        dbplyr_1.4.2      compiler_3.6.1    pillar_1.4.2      prettyunits_1.0.2 remotes_2.1.0
 [7] tools_3.6.1       progress_1.2.2    bit_1.1-14        testthat_2.2.1    zeallot_0.1.0     digest_0.6.22
[13] pkgbuild_1.0.6    pkgload_1.0.2     RSQLite_2.1.2     memoise_1.1.0     tibble_2.1.3      pkgconfig_2.0.3
[19] rlang_0.4.1       rex_1.1.2         cli_1.1.0         DBI_1.0.0         yaml_2.2.0        pkgreviewr_0.1.2
[25] arkdb_0.0.5       withr_2.1.2       rappdirs_0.3.1    desc_1.2.0        fs_1.3.1          vctrs_0.2.0
[31] devtools_2.2.1    hms_0.5.1         bit64_0.9-7       tidyselect_0.2.5  rprojroot_1.3-2   glue_1.3.1
[37] R6_2.4.0          processx_3.4.1    fansi_0.4.0       sessioninfo_1.1.1 blob_1.2.0        purrr_0.3.3
[43] callr_3.3.2       covr_3.3.2        magrittr_1.5      backports_1.1.5   ps_1.3.0          ellipsis_0.3.0
[49] usethis_1.5.1     assertthat_0.2.1  utf8_1.1.4        stringi_1.4.3     lazyeval_0.2.2    crayon_1.3.4
```

On Mac:
```r
Platform: x86_64-apple-darwin15.6.0 (64-bit)
Running under: macOS Mojave 10.14.6

BLAS:   /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib
LAPACK: /Library/Frameworks/R.framework/Versions/3.6/Resources/lib/libRlapack.dylib

[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

[1] devtools_2.2.1 usethis_1.5.1  magrittr_1.5   taxadb_0.0.1   readr_1.3.1    dplyr_0.8.3

 [1] Rcpp_1.0.2         whoami_1.3.0       prettyunits_1.0.2  ps_1.3.0           clisymbols_1.2.0   utf8_1.1.4
 [7] assertthat_0.2.1   zeallot_0.1.0      rprojroot_1.3-2    digest_0.6.22      R6_2.4.0           backports_1.1.5
[13] RSQLite_2.1.2      evaluate_0.14      httr_1.4.1         pillar_1.4.2       rlang_0.4.1        progress_1.2.2
[19] curl_4.2           lazyeval_0.2.2     rstudioapi_0.10    praise_1.0.0       MonetDBLite_0.6.1  pkgreviewr_0.1.3
[25] callr_3.3.2        blob_1.2.0         rmarkdown_1.16     desc_1.2.0         stringr_1.4.0      bit_1.1-14
[31] compiler_3.6.1     lintr_2.0.0        xfun_0.10          pkgconfig_2.0.3    base64enc_0.1-3    pkgbuild_1.0.6
[37] htmltools_0.4.0    tidyselect_0.2.5   tibble_2.1.3       codetools_0.2-16   fansi_0.4.0        crayon_1.3.4
[43] dbplyr_1.4.2       withr_2.1.2        rappdirs_0.3.1     jsonlite_1.6       DBI_1.0.0          covr_3.3.2
[49] cli_1.1.0          stringi_1.4.3      fs_1.3.1           remotes_2.1.0      rex_1.1.2          testthat_2.2.1
[55] xml2_1.2.2         ellipsis_0.3.0     vctrs_0.2.0        cyclocomp_1.1.0    arkdb_0.0.5        tools_3.6.1
[61] rcmdcheck_1.3.3    bit64_0.9-7        glue_1.3.1         purrr_0.3.3        hms_0.5.1          processx_3.4.1
[67] pkgload_1.0.2      yaml_2.2.0         xmlparsedata_1.0.3 xopen_1.0.0        sessioninfo_1.1.1  memoise_1.1.0
[73] goodpractice_1.0.2 knitr_1.25

```

rOpenSci guidelines:
- [X] Does the code comply with general principles in the Mozilla reviewing guide?
- [X] Does the package comply with the rOpenSci packaging guide?
