### 2019-10-29 #344 @karinorman
Hi @lindsayplatt & @mcsiple, thanks for taking the time to review and for your helpful comments! A couple general responses below about issues you both ran into with more detailed and individualized responses to come!

**Vignettes**

Thanks @ldecicco-USGS for pointing out the vignette location. They are currently in the `articles` subdirectory to prevent timing out on CRAN (`pkgdown` still detects and builds vignettes in subdirectories). Prebuilt vignettes are linked at the bottom of the README, but we could also put a lighter weight version in `vingettes/` so they're accessible by `browseVignettes(package="taxadb")`.

**Backend Documentation**

It's clear from both of your experiences with the package that an upfront description of the package backends (i.e. database hosts) is necessary. We will expand on that in the `schema.rmd` vignette, but also wanted to point to the article draft in `paper/manuscript.rmd`, which we would also love feedback on! It lays out the backends in a lot more detail and should lend some clarity.

@lindsayplatt, backend problems are likely the source of your performance issues. `td_create("col")` only installs the database for Catalogue of Life, whereas many of the examples reference other providers (such as ITIS, which is the default provider). If you try to use a provider that has not been installed the default is to install that database on the fly, which takes a while!  Running `td_create("all")` will install all databases ahead of time.   This is obviously an issue of clarity on our part in the README, but just wanted to explain what was likely going on there.

Also, the freezing behavior Lindsay observed is almost surely due to the computer running low on memory.  We suspect this is because `RSQLite` or `MonetDBLite` were not installed?  We made `RSQLite` an optional dependency and allowed `taxadb` to work in-memory if it was not found, but this example demonstrates that was a poor choice, since users can easily run out of memory with these giant data tables!  We've making RSQLite a required dependency now, and adding a vignette to explain how to choose between these.  (Basically RSQLite is easier to install but not as fast as MonetDBLite!)  You can check `sessionInfo()` or just run `taxadb::td_connect()` to see what database connection is being used.
