### 2019-10-30 c863afc
Import RSQLite and clarify the multi-db methods
