+++ vignettes/articles/data-sources.Rmd
@@ -1,0 +1,109 @@
+---
+title: "Name Providers and schema used in taxadb"
+author: "Carl Boettiger"
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{schema}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r setup, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+taxadb:::td_disconnect()
+MonetDBLite::monetdblite_shutdown()
+```
+
+
+
+`taxadb` relies on a set of pre-assembled tables following a set of standardized schema layouts using Darwin Core vocabulary, as outlined below.  The database dumps provided by providers supported in `taxadb` at this time are:
+
+`taxadb` abbreviation | name | update frequency | synthetic
+
+`itis`   | [The Integrated Taxonomic Information System](https://www.itis.gov/)
+`col`  | [The Catalogue of Life](http://www.catalogueoflife.org/)
+`ncbi` |  [The National Center for Biotechnology Information](https://www.ncbi.nlm.nih.gov/)
+`gbif` | [The Global Biodiversity Information Facility](https://www.gbif.org/)
+`tpl`  | [The Plant List](http://www.theplantlist.org/)
+`fb`   | [FishBase](https://fishbase.ca)
+`slb`  | [SeaLifeBase](https://www.sealifebase.ca/)
+`wd`   | [WikiData](https://wikidata.org)
+`iucn` | [The IUCN Red List](https://www.iucnredlist.org) of endangered species status
+`ott`  | [Open Tree of Life taxonomy](https://tree.opentreeoflife.org/about/taxonomy-version/ott3.1).  
+
+
+***Please Note***: `taxadb` advises against uncritically combining data from multiple providers.  The same name is frequently used by different providers to mean different things -- some providers consider two names synonyms that other providers consider distinct species.  *It is crucial to recognize that taxonomic name providers represent independent taxonomic theories*, and not merely additional observations of the same immutable reality ([Franz & Sterner (2018)](https://doi.org/10.1093/database/bax100 "Nico M Franz, Beckett W Sterner, To increase trust, change the social design behind aggregated biodiversity data, Database, Volume 2018, 2018, bax100")). You cannot just merge two databases of taxonomic names like you can two databases of, say, plant traits to get a bigger and more complete sample, because the former can contain meaningful controdictions.  
+
+
+At the same time, it is also important to note that `col`, `gbif`, `ott`, are explicitly synthesis projects integrating the databases of names from a range of (many) other providers, while `itis`, `iucn`, `ncbi`, `tpl`, `fb`, and `slb` are independent name providers.  The synthetic or integrated name lists are not simple merges, but the product of considerably expert opinion, and occassional nonsense automation. As such, they too represent novel (justified or otherwise) assertions of taxonomy, and are in no way a complete substitute for the databases they integrate, owing to both differences in how up-to-date the relative records are as well as to either expert disagreements or algorithmic miss-matches.  `taxadb` makes no attempt to provide an opinion or reconciliation mechanism to any of these issues, but only to provide convenient access to data and functions for manipulating these records in a fast and consistent manner.  (In fact, it is easy to use `taxadb` to verify that many of the names recognized in, say, ITIS, are not in fact included at all in Catalogue of Life or other databases that claim to derive from ITIS).
+
[…]
