+++ README.md
@@ -97,1509 +97,15 @@ help us resolve this last case (see below).
 
 ``` r
 birds %>% 
-  mutate(accepted_name = get_names(id, "col"))
-#>                                           species           id
-#> 1                          Dendrocygna autumnalis         <NA>
-#> 2                             Dendrocygna bicolor COL:35517332
-#> 3                                 Anser canagicus COL:35517329
-#> 4                              Anser caerulescens COL:35517325
-#> 5                   Chen caerulescens (blue form)         <NA>
-#> 6                                    Anser rossii COL:35517328
-#> 7                                 Anser albifrons         <NA>
-#> 8                                 Branta bernicla COL:35517301
-#> 9                       Branta bernicla nigricans COL:35537100
-#> 10                              Branta hutchinsii COL:35536445
-#> 11                              Branta canadensis COL:35517289
-#> 12                                    Cygnus olor COL:35517278
-#> 13                              Cygnus buccinator         <NA>
-#> 14                             Cygnus columbianus         <NA>
-#> 15                               Cairina moschata COL:35517495
-#> 16                                     Aix sponsa         <NA>
-#> 17                                Spatula discors         <NA>
-#> 18                             Spatula cyanoptera         <NA>
-#> 19                               Spatula clypeata COL:42095596
-#> 20                                Mareca strepera         <NA>
-#> 21                                Mareca penelope         <NA>
-#> 22                               Mareca americana COL:35517378
-#> 23                             Anas platyrhynchos COL:35517347
-#> 24                       Anas platyrhynchos diazi COL:35537080
-#> 25                                  Anas rubripes COL:35517352
-#> 26                                 Anas fulvigula COL:35517354
-#> 27     Anas platyrhynchos x rubripes or fulvigula         <NA>
-#> 28                                     Anas acuta COL:35517358
-#> 29                                    Anas crecca COL:35517365
-#> 30                             Aythya valisineria         <NA>
-#> 31                               Aythya americana COL:35517406
-#> 32                                Aythya collaris COL:35517409
-#> 33                                Aythya fuligula COL:35517416
-#> 34                                  Aythya marila COL:35517411
-#> 35                                 Aythya affinis         <NA>
-#> 36                        Aythya marila / affinis         <NA>
-#> 37                          Somateria spectabilis COL:35517435
-#> 38                           Somateria mollissima COL:35517430
-#> 39                      Histrionicus histrionicus COL:35517427
-#> 40                        Melanitta perspicillata COL:35517444
-#> 41                                Melanitta fusca         <NA>
-#> 42                            Melanitta americana COL:35517447
-#> 43    Melanitta perspicillata / fusca / americana         <NA>
-#> 44                              Clangula hyemalis COL:35517426
-#> 45                              Bucephala albeola COL:35517425
-#> 46                             Bucephala clangula COL:35517421
-#> 47                            Bucephala islandica COL:35517424
-#> 48                 Bucephala clangula / islandica         <NA>
-#> 49                          Lophodytes cucullatus COL:35517455
-#> 50                               Mergus merganser COL:35517456
-#> 51                                Mergus serrator COL:35517458
-#> 52                             Oxyura jamaicensis COL:35517448
-#> 53                                 Ortalis vetula COL:35517860
-#> 54                                Oreortyx pictus COL:35518013
-#> 55                            Colinus virginianus         <NA>
-#> 56                            Callipepla squamata COL:35517994
-#> 57                         Callipepla californica COL:35517998
-#> 58                            Callipepla gambelii COL:35517999
-#> 59                            Cyrtonyx montezumae COL:35518019
-#> 60                               Alectoris chukar COL:35518025
[…]
