### 2019-10-30 fc4978c
update manuscript and docs
