+++ vignettes/articles/data-sources.Rmd
@@ -1,6 +1,6 @@
 ---
 title: "Name Providers and schema used in taxadb"
-author: "Carl Boettiger"
+author: "Carl Boettiger, Kari Norman"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
@@ -14,16 +14,14 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
-taxadb:::td_disconnect()
-MonetDBLite::monetdblite_shutdown()
 ```
 
 
 
 `taxadb` relies on a set of pre-assembled tables following a set of standardized schema layouts using Darwin Core vocabulary, as outlined below.  The database dumps provided by providers supported in `taxadb` at this time are:
 
-`taxadb` abbreviation | name | update frequency | synthetic
-
+`taxadb` abbreviation | name 
+----------------------|-------------------
 `itis`   | [The Integrated Taxonomic Information System](https://www.itis.gov/)
 `col`  | [The Catalogue of Life](http://www.catalogueoflife.org/)
 `ncbi` |  [The National Center for Biotechnology Information](https://www.ncbi.nlm.nih.gov/)
@@ -47,14 +45,7 @@ Most common operations can be expressed in terms of standard database operations
 
 ## Data Schema
 
-`taxadb` relies on the Simple Darwin Core Namespace for Taxon objects, <http://rs.tdwg.org/dwc/terms/> [@dwc].  This is the mostly widely recognized format for exchange of taxonomic information
-
-```{r message = FALSE}
-library(dplyr)
-library(taxadb)
-td_create("all")
-```
-
+`taxadb` relies on the Simple Darwin Core Namespace for Taxon objects, <http://rs.tdwg.org/dwc/terms/> [@dwc].  This is the mostly widely recognized format for exchange of taxonomic information.
 
 - `taxonID`: a unique id for the name (including provider prefix).  Note that some providers do not assign IDs to synonyms, but only to accepted names.  In this case, the `taxonID` should be `NA`, and the ID to the accepted name should be specified in `acceptedNameUsageID`.  
 - `scientificName`: a Latin name, either accepted or known synonym, at the lowest resolved level for the taxon.  While DWC encourages the use of authorship citations, these are intentionally omitted in most tables as inconsistency in abbreviations and formatting make names with authors much harder to resolve.  When available, this information is provided in the additional optional columns using the corresponding Darwin Core terms.  ***Please note***: `scientificName` includes names at all taxonomic rank levels, it does not mean just "genus + specific epithet".  For example, "Animalia" is also a scientific name.  The `taxonRank` column indicates the associated taxonomic rank.  
@@ -102,8 +93,3 @@ Common names are available from several providers, but tidy tables for `taxadb` 
   and rank names should have 
 - Encoding is UTF-8
 
-
-
-```{r include=FALSE}
-taxadb:::td_disconnect()
-```
