### 2019-10-30 35eb39b
Make test more comprehensive.
