### 2019-10-30 a3e03c7
Rename functions changing by_* to filter_*
