### 2019-10-30 398e229
Update other functions' documentation and use of filter_ functions.
