### 2019-10-30 9017c76
adapt tests
