### 2019-10-30 1a871ee
Update examples with current function names.
