+++ R/td_connect.R
@@ -31,7 +31,7 @@
 #' Sys.setenv(TAXADB_HOME=tempdir())
 #'
 #' ## Connect to the database:
-#' db <- connect_db()
+#' db <- td_connect()
 #'
 #' }
 td_connect <- function(dbdir = taxadb_dir(),
