### 2019-11-05 #344 @karinorman
@ldecicco-USGS We're also still working on a few of the comments made by both reviewers, including examples and test for a couple functions, better common names examples, etc. Hoping to have that finished up in the next couple days but would love responses to the above in the meantime. Thanks everyone!
