### 2019-11-05 1dde53f
Modify to allow for arguments passed to the operating function
