### 2019-11-06 5aebf83
Add option to remove punctuation, for use with common names.
