### 2019-11-06 #344 @mcsiple
Hi all -- apologies for the delayed response. Thank you for adding the extended documentation.

- The added database backends vignette is great! Thank you for that.

- I quite like the extended vignette you added @cboettig . The only potential issue I ran into was that the current chunk inside `system.time({})`  took a while for me to run. If other people besides me report really long running times, it might be worth finding an example with a shorter run time. Mine completed in the following:
```{r}
   user  system elapsed
 448.79  227.97  686.84
```
Which is much longer than it says the example took in the vignette. A second attempt took about the same amount of time. Besides that it is great and I liked the added length/detail.

- Just to clarify after looking over the manuscript draft - is `taxadb` intended to replace `taxize` ? If not, maybe it's worth discussing in the ms how they can be used together. I am imagining a user who is brand new to taxonomic databases and wants to use the most rigorous workflow to match a list of species to these databases and extract some info about IUCN or CITES listing (for example).
