### 2019-11-06 bda3010
little documentation things to get rid of warnings
