+++ R/mutate_db.R
@@ -25,11 +25,13 @@
 #'   }
 #'
 #'   #Clean a list of messy common names
-#'   names <- clean_names(c("Steller's jay", "coopers Hawk"), binomial_only = FALSE, remove_sp = FALSE, remove_punc = TRUE)
+#'   names <- clean_names(c("Steller's jay", "coopers Hawk"),
+#'                binomial_only = FALSE, remove_sp = FALSE, remove_punc = TRUE)
 #'
 #'   #Get cleaned common names from a provider and search for cleaned names in that table
 #'   taxa_tbl("itis", "common") %>%
-#'   mutate_db(clean_names, "vernacularName", "vernacularNameClean", binomial_only = FALSE, remove_sp = FALSE, remove_punc = TRUE) %>%
+#'   mutate_db(clean_names, "vernacularName", "vernacularNameClean",
+#'             binomial_only = FALSE, remove_sp = FALSE, remove_punc = TRUE) %>%
 #'   filter(vernacularNameClean %in% names)
 #'
 #'
