+++ NAMESPACE
@@ -67,7 +67,6 @@ importFrom(stringi,stri_trans_general)
 importFrom(stringi,stri_trim)
 importFrom(tibble,as_tibble)
 importFrom(tibble,column_to_rownames)
-importFrom(tibble,lst)
 importFrom(tibble,tibble)
 importFrom(utils,download.file)
 importFrom(utils,head)
