### 2019-11-07 57b733c
Export td_disconnect, which closes a connection to the database.
