+++ NAMESPACE
@@ -18,6 +18,7 @@ export(synonyms)
 export(taxa_tbl)
 export(td_connect)
 export(td_create)
+export(td_disconnect)
 importFrom(DBI,dbClearResult)
 importFrom(DBI,dbConnect)
 importFrom(DBI,dbCreateTable)
