### 2019-11-20 8844e13
Revised data-raw scripts
- Synonyms were missing hierarchy in col
- ITIS had lost synonyms
- GBIF tidy
- sealifebase fixed
- IUCN revision in progress
