+++ data-raw/fishbase.R
@@ -3,249 +3,238 @@ library(tidyverse)
 library(stringi)
 source("data-raw/helper-routines.R")
 
-#### Fishbase
-fb <- as_tibble(rfishbase::load_taxa())
-fb_wide <- fb %>%
-  select( id = SpecCode,
-          species = Species,
-          genus = Genus,
-          subfamily = Subfamily,
-          family = Family,
-          order = Order,
-          class = Class,
-          superclass = SuperClass) %>%
-  mutate(phylum = "Chorodata",
-         kingdom = "Animalia",
-         id = stri_paste("FB:", id))
+preprocess_fb <- function(){
+  #### Fishbase
+  fb <- as_tibble(rfishbase::load_taxa())
+  fb_wide <- fb %>%
+    select( id = SpecCode,
+            species = Species,
+            genus = Genus,
+            subfamily = Subfamily,
+            family = Family,
+            order = Order,
+            class = Class,
+            superclass = SuperClass) %>%
+    mutate(phylum = "Chorodata",
+           kingdom = "Animalia",
+           id = stri_paste("FB:", id))
 
-accepted <- fb_wide %>% select(id, species) %>% gather(rank, name, -id)
-species <- rfishbase:::fb_species()
-synonym_table <- rfishbase::synonyms(NULL) %>%
-  left_join(species) %>%
-  rename(id = SpecCode)  %>%
-  mutate(id = stri_paste("FB:", id),
-         SynCode = stri_paste("FB:", SynCode),
-         TaxonLevel = tolower(TaxonLevel)) %>%
-  select(id,
-         accepted_name = Species,
-         name = synonym,
-         type = Status,
-         synonym_id = SynCode,
-         rank = TaxonLevel)
+  accepted <- fb_wide %>% select(id, species) %>% gather(rank, name, -id)
+  species <- rfishbase:::fb_species()
+  synonym_table <- rfishbase::synonyms(NULL) %>%
+    left_join(species) %>%
+    rename(id = SpecCode)  %>%
+    mutate(id = stri_paste("FB:", id),
+           SynCode = stri_paste("FB:", SynCode),
+           TaxonLevel = tolower(TaxonLevel)) %>%
+    select(id,
+           accepted_name = Species,
+           name = synonym,
+           type = Status,
+           synonym_id = SynCode,
+           rank = TaxonLevel)
 
-#find accepted names that are actually useful synonym maps, create table with just synonyms
-synonyms <- synonym_table %>%
-  filter(type == "accepted name", accepted_name != name) %>%
-  select(-type) %>%
-  mutate(type = "epithet synonym") %>%
-  rename(name = "accepted_name", accepted_name = "name") %>%
-  bind_rows(synonym_table %>% filter(type != "accepted name"))
+  #find accepted names that are actually useful synonym maps, create table with just synonyms
+  synonyms <- synonym_table %>%
+    filter(type == "accepted name", accepted_name != name) %>%
+    select(-type) %>%
+    mutate(type = "epithet synonym") %>%
+    rename(name = "accepted_name", accepted_name = "name") %>%
+    bind_rows(synonym_table %>% filter(type != "accepted name"))
 
-## Merge taxon_id and synonym id
-fb_taxonid <- synonyms %>%
-  rename(accepted_id = id, id = synonym_id) %>%
-  select(id, name, rank, accepted_id, name_type = type) %>%
-  bind_rows(mutate(accepted, accepted_id = id, name_type = "accepted")) %>%
-  distinct()  %>% de_duplicate()
+  ## Merge taxon_id and synonym id
[…]
