+++ data-raw/col.R
@@ -1,92 +1,111 @@
 library(dplyr)
 library(readr)
+library(forcats)
 library(stringi)
 source("data-raw/helper-routines.R")
 
 # All snapshots available from: http://www.catalogueoflife.org/DCA_Export/archive.php
-dir.create("taxizedb/col", FALSE, TRUE)
-download.file("http://www.catalogueoflife.org/DCA_Export/zip-fixed/2018-annual.zip",
-              "taxizedb/col/2018-annual.zip")
-unzip("taxizedb/col/2018-annual.zip", exdir="taxizedb/col")
-taxon <- read_tsv("taxizedb/col/taxa.txt", col_types = cols(.default = col_character()), quote = "")
-vernacular <- read_tsv("taxizedb/col/vernacular.txt", col_types = cols(.default = col_character()), quote = "")
-reference <- read_tsv("taxizedb/col/reference.txt", col_types = cols(.default = col_character()), quote = "")
 
-## Not useful to us, also not very complete:
-## distribution <- read_tsv("taxizedb/col/distribution.txt", col_types = cols(.default = col_character()), quote = "")     # occurrance, but sparse & coarse
-## speciesprofile <- read_tsv("taxizedb/col/speciesprofile.txt", col_types = cols(.default = col_character()), quote = "") # habitat
-# description <- read_tsv("taxizedb/col/description.txt", col_types = cols(.default = col_character()), quote = "") ## occurance countries
+preprocess_col <- function(year = "2019", dir = file.path(tempdir(), "col")){
 
+  dir.create(dir, FALSE, FALSE)
+  download.file(paste0("http://www.catalogueoflife.org/DCA_Export/zip-fixed/",
+                       year, "-annual.zip"),
+                file.path(dir, "col-annual.zip"))
+  unzip(file.path(dir, "col-annual.zip"), exdir=dir)
 
-## scientificNameAuthorship tagged on to scientificName, and in inconsistent format. trim it off.
-taxa <- taxon %>%
-  mutate(taxonomicStatus = forcats::fct_recode(taxonomicStatus, "accepted" = "accepted name")) %>%
-  select(taxonID, scientificName, acceptedNameUsageID, taxonomicStatus, taxonRank,
-         kingdom, phylum, class, order, family, genus, specificEpithet, infraspecificEpithet,
-         taxonConceptID, isExtinct, nameAccordingTo, namePublishedIn, scientificNameAuthorship)
-
-taxa <- bind_rows(
-  taxa %>%
-    filter(!is.na(scientificNameAuthorship)) %>%
-    mutate(scientificName =
-           stri_trim(stri_replace_first_fixed(scientificName, scientificNameAuthorship, ""))),
-  taxa %>%
-    filter(is.na(scientificNameAuthorship))
-)
+  ## a better read_tsv
+  read_tsv <- function(...) readr::read_tsv(..., quote = "", col_types = readr::cols(.default = "c"))
 
 
+  taxon <- read_tsv(file.path(dir, "taxa.txt"))
+  reference <- read_tsv(file.path(dir, "reference.txt"))
 
-## For accepted names, set acceptedNameUsageID to match taxonID, rather NA
-accepted <-
-  filter(taxa, taxonomicStatus == "accepted") %>%
-  mutate(acceptedNameUsageID = taxonID)
-rest <-
-  filter(taxa, taxonomicStatus != "accepted") %>%
-  filter(!is.na(acceptedNameUsageID)) # We drop un-mapped synonyms, as they are not helpful
[…]

[…]
