### 2019-11-20 ff3bb70
Add code of conduct
