+++ README.Rmd
@@ -168,3 +168,8 @@ MonetDBLite::monetdblite_shutdown()
 
 if(require(codemetar)) codemetar::write_codemeta()
 ```
+
+----
+
+Please note that this project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md).
+By participating in this project you agree to abide by its terms.
