+++ data-raw/gbif.R
@@ -5,85 +5,88 @@ source("data-raw/helper-routines.R")
 
 ## extracted from: https://doi.org/10.15468/39omei
 
-downloads <- tempdir()
-dir <- file.path(downloads, "gbif")
-dir.create(dir, FALSE, FALSE)
-download.file("http://rs.gbif.org/datasets/backbone/backbone-current.zip",
-              file.path(dir, "backbone.zip"))
-unzip(file.path(dir, "backbone.zip"), exdir=dir)
+  preprocess_gbif <- function(year = "2019", dir = file.path(tempdir(), "gbif")){
 
-## a better read_tsv
-read_tsv <- function(...) readr::read_tsv(..., quote = "", col_types = readr::cols(.default = "c"))
+  dir.create(dir, FALSE, FALSE)
+  download.file("http://rs.gbif.org/datasets/backbone/backbone-current.zip",
+                file.path(dir, "backbone.zip"))
+  unzip(file.path(dir, "backbone.zip"), exdir=dir)
 
-taxon <- read_tsv(file.path(dir, "Taxon.tsv"))
+  ## a better read_tsv
+  read_tsv <- function(...) readr::read_tsv(..., quote = "", col_types = readr::cols(.default = "c"))
 
+  ## And here we go!
+  taxon <- read_tsv(file.path(dir, "Taxon.tsv"))
+  ## canonicalName appears to be: Genus + specificEpithet + infraspecificEpithet
+  ## i.e. SpecificEpithet ~ a name at the "species" rank
+  ## Scientific name is ~ canonical name + citation parenthetical
+  ## Darwin Core defines ScientificName as: 	The full scientific name,
+  ## with authorship and date information if known. When forming part of an
+  ## Identification, this should be the name in lowest level taxonomic rank
+  ## that can be determined. This term should not contain identification
+  ## qualifications, which should instead be supplied in the IdentificationQualifier term.
 
+  ## This is unfortunate as author & date formatting introduces a huge
+  ## amount of variation that is difficult to resolve against, and best treated as a separate field.
 
-## canonicalName appears to be: Genus + specificEpithet + infraspecificEpithet
-## i.e. SpecificEpithet ~ a name at the "species" rank
-## Scientific name is ~ canonical name + citation parenthetical
-## Darwin Core defines ScientificName as: 	The full scientific name,
-## with authorship and date information if known. When forming part of an
-## Identification, this should be the name in lowest level taxonomic rank
-## that can be determined. This term should not contain identification
-## qualifications, which should instead be supplied in the IdentificationQualifier term.
+  ## genericName and canonicalName are not Darwin Core Taxon properties
+  ## Note that even within GBIF, formatting of scientificNameAuthorship
+  ## is highly non-standard (parens, abbrv, initials, etc)
+  gbif <- taxon %>%
+    select(taxonID,
+           scientificName = canonicalName,
+           taxonRank,
+           taxonomicStatus,
+           acceptedNameUsageID,
+           kingdom, phylum, class, order, family, genus, specificEpithet, infraspecificEpithet,
+           parentNameUsageID,
+           originalNameUsageID,
+           scientificNameAuthorship)
 
-## This is unfortunate as author & date formatting introduces a huge
-## amount of variation that is difficult to resolve against, and best treated as a separate field.
-
[…]
