### 2019-12-02 #344 @karinorman
@ldecicco-USGS Yes! Just wanted to address some of the general changes we've now made based on comments from both reviewers.

- contribution guidelines added
- changed naming convention from `by_` to `filter_` to reflect the intuition that these functions filter underlying tables based on specified input
- updated examples and function tests, we now have 94% coverage based on [codev](https://codecov.io/gh/cboettig/taxadb)
- added function `td_disconnect()` to manually close the database connection
- added more detailed examples for working with messy common names, for example see `mutate_db` documentation

We are currently still working on a way to deal with database versioning so that multiple versions can be installed on a single machine, with the ability to move between versions. We plan to have that finished up in the next couple weeks.

@mcsiple and @lindsayplatt, let us know if there is anything else you would like to discuss further and thanks again for your helpful feedback!
