### 2019-12-02 77cc926
still streamlining. NCBI runs now
