+++ data-raw/itis.R
@@ -2,209 +2,222 @@ library(tidyverse)
 library(stringi)
 library(piggyback)
 library(RSQLite)
+library(openssl)
 source("data-raw/helper-routines.R")
 
 
+preprocess_itis <- function(url = "https://www.itis.gov/downloads/itisSqlite.zip",
+                            output_paths =
+                              c(dwc = "2019/dwc_itis.tsv.bz2",
+                                common = "2019/common_itis.tsv.bz2")
+){
+
+  archive <- file.path(tempdir(), "itis",  "itisSqlite.zip")
+  dir <- dirname(archive)
+  dir.create(dir, FALSE, FALSE)
+  download.file(url, archive)
+  message(paste(openssl::sha256(archive)))
+
+  unzip(archive, exdir=dir)
+  dbname <- list.files(list.dirs(dir, recursive=FALSE), pattern="[.]sqlite", full.names = TRUE)
+  db <- DBI::dbConnect(RSQLite::SQLite(), dbname = dbname)
+  arkdb::ark(db, dir, arkdb::streamable_readr_tsv(), lines = 1e6L)
+
+  ## defaulting to logicals is so annoying!
+  read_tsv <- function(...) readr::read_tsv(..., quote = "", col_types = readr::cols(.default = "c"))
 
 
-downloads <- tempdir()
-dir <- file.path(downloads, "itis")
-dir.create(dir, FALSE, FALSE)
-itis_sql <- file.path(dir,  "itisSqlite.zip")
-#### ITIS DIRECT:
+  ## not that rank_id isn't a unique id by itself!
+  rank_tbl <-
+    read_tsv(file.path(dir, "taxon_unit_types.tsv.bz2")) %>%
+    select(kingdom_id, rank_id, rank_name) %>%
+    collect() %>%
+    unite(rank_id, -rank_name, sep = "-") %>%
+    mutate(rank_name =
+             stringr::str_remove_all(
+              stringr::str_to_lower(rank_name),"\\s"))
 
-download.file("https://www.itis.gov/downloads/itisSqlite.zip", itis_sql)
-unzip(itis_sql, exdir=dir)
-dbname <- list.files(list.dirs(dir, recursive=FALSE), pattern="[.]sqlite", full.names = TRUE)
-db <- DBI::dbConnect(RSQLite::SQLite(), dbname = dbname)
-arkdb::ark(db, dir, arkdb::streamable_readr_tsv(), lines = 1e6L)
+  hierarch <-
+    read_tsv(file.path(dir, "taxonomic_units.tsv.bz2")) %>%
+    mutate(rank_id = paste(kingdom_id, rank_id, sep="-")) %>%
+    select(tsn, parent_tsn, rank_id, complete_name) %>% distinct()
 
-## defaulting to logicals is so annoying!
-read_tsv <- function(...) readr::read_tsv(..., col_types = readr::cols(.default = "c"))
-
-
-## not that rank_id isn't a unique id by itself!
-rank_tbl <-
-  read_tsv(file.path(dir, "taxon_unit_types.tsv.bz2")) %>%
-  select(kingdom_id, rank_id, rank_name) %>%
-  collect() %>%
-  unite(rank_id, -rank_name, sep = "-") %>%
-  mutate(rank_name =
-           stringr::str_remove_all(
-            stringr::str_to_lower(rank_name),"\\s"))
-
-hierarch <-
-  read_tsv(file.path(dir, "taxonomic_units.tsv.bz2")) %>%
-  mutate(rank_id = paste(kingdom_id, rank_id, sep="-")) %>%
-  select(tsn, parent_tsn, rank_id, complete_name) %>% distinct()
-
[…]
