### 2019-12-02 1b379d8
use file hash, not string hash!
