+++ data-raw/make.R
@@ -37,17 +37,32 @@ itis = preprocess_itis(url = file_in("https://www.itis.gov/downloads/itisSqlite.
                        output_paths = c(dwc = file_out("2019/dwc_itis.tsv.bz2"),
                                         common = file_out("2019/common_itis.tsv.bz2")))
 
-ncbi = preprocess_ncbi(url = "ftp://ftp.ncbi.nih.gov/pub/taxonomy/taxdmp.zip",
-                      output_paths = c(dwc = "2019/dwc_ncbi.tsv.bz2",
-                                       common = "2019/common_ncbi.tsv.bz2"))
+ncbi = preprocess_ncbi(url = file_in("ftp://ftp.ncbi.nih.gov/pub/taxonomy/taxdmp.zip"),
+                      output_paths = c(dwc = file_out("2019/dwc_ncbi.tsv.bz2"),
+                                       common = file_out("2019/common_ncbi.tsv.bz2")))
+
+col = preprocess_col(url = file_in("http://www.catalogueoflife.org/DCA_Export/zip-fixed/2019-annual.zip"),
+                     output_paths = c(dwc = file_out("2019/dwc_col.tsv.bz2"),
+                                      common = file_out("2019/common_col.tsv.bz2")))
 
 
-)
+fb = preprocess_fb(output_paths = c(dwc = file_out("2019/dwc_fb.tsv.bz2"),
+                                    common = file_out("2019/common_fb.tsv.bz2")))
+
+slb = preprocess_slb(output_paths = c(dwc = file_out("2019/dwc_slb.tsv.bz2"),
+                                    common = file_out("2019/common_slb.tsv.bz2")))
+
+ott = preprocess_ott(url = file_in("http://files.opentreeoflife.org/ott/ott3.2/ott3.2.tgz"),
+                     output_paths = c(dwc = file_out("2019/dwc_ott.tsv.bz2"),
+                                      common = file_out("2019/common_ott.tsv.bz2")))
 
 
-config <- drake_config(plan)
-vis_drake_graph(config)
-drake::make(plan)
+#)
+
+
+#config <- drake_config(plan)
+#vis_drake_graph(config)
+#drake::make(plan)
 
 
 # piggyback::pb_upload( "dwc/fb.tsv.bz2", repo="boettiger-lab/taxadb-cache", tag = "dwc")
