### 2019-12-10 322b18e
Citations by hand in the provider table.
