### 2019-12-11 731f1bf
Beef up citations
