+++ paper/refs.bib
@@ -46,7 +46,7 @@
 }
 @misc{ncbi,
 	title = {NCBI {Taxonomy}},
-	url = {https://www.itis.gov},
+	url = {https://www.ncbi.nlm.nih.gov/taxonomy},
 	author = {National Center for Biotechnology Information},
 	publisher = {U.S. National Library of Medicine},
   address = {8600 Rockville Pike, Bethesda MD, 20894 USA},
@@ -206,4 +206,166 @@ and Logs},
 	url = {https://www.gbif.org},
 	author = {GBIF},
 	year = {2019}
+}
+
+@article{boyle2013,
+	title = {The taxonomic name resolution service: an online tool for automated standardization of plant names},
+	volume = {14},
+	issn = {1471-2105},
+	shorttitle = {The taxonomic name resolution service},
+	url = {https://doi.org/10.1186/1471-2105-14-16},
+	doi = {10.1186/1471-2105-14-16},
+	abstract = {The digitization of biodiversity data is leading to the widespread application of taxon names that are superfluous, ambiguous or incorrect, resulting in mismatched records and inflated species numbers. The ultimate consequences of misspelled names and bad taxonomy are erroneous scientific conclusions and faulty policy decisions. The lack of tools for correcting this ‘names problem’ has become a fundamental obstacle to integrating disparate data sources and advancing the progress of biodiversity science.},
+	number = {1},
+	urldate = {2019-12-11},
+	journal = {BMC Bioinformatics},
+	author = {Boyle, Brad and Hopkins, Nicole and Lu, Zhenyuan and Raygoza Garay, Juan Antonio and Mozzherin, Dmitry and Rees, Tony and Matasci, Naim and Narro, Martha L. and Piel, William H. and Mckay, Sheldon J. and Lowry, Sonya and Freeland, Chris and Peet, Robert K. and Enquist, Brian J.},
+	month = jan,
+	year = {2013},
+	pages = {16},
+	file = {Boyle et al_2013_BMC Bioinformatics.pdf:/Users/karinorman/Dropbox/Literature/Boyle et al_2013_BMC Bioinformatics.pdf:application/pdf}
+}
+
+@article{dayrat2005,
+	title = {Towards integrative taxonomy},
+	volume = {85},
+	issn = {0024-4066},
+	url = {https://academic.oup.com/biolinnean/article/85/3/407/2691659},
+	doi = {10.1111/j.1095-8312.2005.00503.x},
+	abstract = {Abstract.  Delineating species boundaries correctly is crucial to the discovery of life's diversity because it determines whether or not different individual or},
+	language = {en},
+	number = {3},
+	urldate = {2019-12-11},
+	journal = {Biological Journal of the Linnean Society},
+	author = {Dayrat, Benoît},
+	month = jul,
+	year = {2005},
+	pages = {407--417},
+	file = {Dayrat_2005_Biological Journal of the Linnean Society.pdf:/Users/karinorman/Dropbox/Literature/Dayrat_2005_Biological Journal of the Linnean Society.pdf:application/pdf}
+}
+
+@article{franz2009,
+	title = {Perspectives: {Towards} a language for mapping relationships among taxonomic concepts},
+	volume = {7},
+	issn = {1477-2000},
+	shorttitle = {Perspectives},
+	url = {https://doi.org/10.1017/S147720000800282X},
+	doi = {10.1017/S147720000800282X},
+	number = {1},
+	urldate = {2019-12-11},
[…]
