### 2019-12-12 b41fcc4
WIP on versioning for the db
