+++ data-raw/helper-routines.R
@@ -33,8 +33,8 @@ n_in_group <- function(data, group_var, ...){
 }
 
 
-file_hash <- function(x, method = openssl::sha256){
-  con <- lapply(x, file)
+file_hash <- function(x, method = openssl::sha256, ...){
+  con <- lapply(x, file, ...)
   hash <- lapply(con, method)
   unlist(lapply(hash, as.character))
 }
