### 2019-12-12 538ea98
regenerate roxygen docs
