+++ man/filter_name.Rd
@@ -4,9 +4,13 @@
 \alias{filter_name}
 \title{Look up taxonomic information by scientific name}
 \usage{
-filter_name(name, provider = c("itis", "ncbi", "col", "tpl", "gbif",
-  "fb", "slb", "wd", "ott", "iucn"), collect = TRUE,
-  ignore_case = TRUE, db = td_connect())
+filter_name(
+  name,
+  provider = c("itis", "ncbi", "col", "tpl", "gbif", "fb", "slb", "wd", "ott", "iucn"),
+  collect = TRUE,
+  ignore_case = TRUE,
+  db = td_connect()
+)
 }
 \arguments{
 \item{name}{a character vector of scientific names, e.g. "Homo sapiens"}
@@ -57,8 +61,10 @@ filter_name(sp)
 
 }
 \seealso{
-Other filter_by: \code{\link{filter_by}},
-  \code{\link{filter_common}}, \code{\link{filter_id}},
-  \code{\link{filter_rank}}
+Other filter_by: 
+\code{\link{filter_by}()},
+\code{\link{filter_common}()},
+\code{\link{filter_id}()},
+\code{\link{filter_rank}()}
 }
 \concept{filter_by}
