+++ man/get_ids.Rd
@@ -4,8 +4,14 @@
 \alias{get_ids}
 \title{get_ids}
 \usage{
-get_ids(names, db = known_providers, format = c("prefix", "bare",
-  "uri"), taxadb_db = td_connect(), ignore_case = TRUE, ...)
+get_ids(
+  names,
+  db = known_providers,
+  format = c("prefix", "bare", "uri"),
+  taxadb_db = td_connect(),
+  ignore_case = TRUE,
+  ...
+)
 }
 \arguments{
 \item{names}{a list of scientific names (which may
@@ -18,10 +24,10 @@ include higher-order ranks in most authorities).}
 \item \code{prefix} (e.g. \code{NCBI:9606}, the default), or
 \item \code{bare} (e.g. \code{9606}, used in \code{taxize::get_ids()}),
 \item \code{uri} (e.g.
-\code{http://ncbi.nlm.nih.gov/taxonomy/9606}).
+\verb{http://ncbi.nlm.nih.gov/taxonomy/9606}).
 }}
 
-\item{taxadb_db}{Connection to from \code{[td_connect]()}.}
+\item{taxadb_db}{Connection to from \verb{[td_connect]()}.}
 
 \item{ignore_case}{should we ignore case (capitalization) in matching names?
 default is \code{TRUE}.}
@@ -35,17 +41,17 @@ To resolve multi-matched names, use \code{\link[=filter_name]{filter_name()}} in
 a table with a separate row for each separate match of the input name.
 }
 \description{
-A drop-in replacement for \code{[taxize::get_ids()]}
+A drop-in replacement for \verb{[taxize::get_ids()]}
 }
 \details{
 Note that some taxize authorities: \code{nbn}, \code{tropicos}, and \code{eol},
 are not recognized by taxadb and will throw an error here. Meanwhile,
-taxadb recognizes several authorities not known to \code{[taxize::get_ids()]}.
+taxadb recognizes several authorities not known to \verb{[taxize::get_ids()]}.
 Both include \code{itis}, \code{ncbi}, \code{col}, and \code{gbif}.
 
 Like all taxadb functions, this function will run
 fastest if a local copy of the provider is installed in advance
-using \code{[td_create()]}.
+using \verb{[td_create()]}.
 }
 \examples{
 \donttest{
@@ -57,6 +63,7 @@ get_ids("Homo sapiens", db= "ncbi", format = "uri")
 \seealso{
 filter_name
 
-Other get: \code{\link{get_names}}
+Other get: 
+\code{\link{get_names}()}
 }
 \concept{get}
