+++ DESCRIPTION
@@ -36,7 +36,7 @@ Imports:
     dbplyr,
     RSQLite
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 6.1.1
+RoxygenNote: 7.0.0
 Suggests:
     spelling,
     testthat,
