+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+export(available_releases)
 export(clean_names)
 export(common_contains)
 export(common_starts_with)
@@ -31,6 +32,7 @@ importFrom(DBI,dbSendQuery)
 importFrom(DBI,dbWriteTable)
 importFrom(arkdb,streamable_readr_tsv)
 importFrom(arkdb,unark)
+importFrom(curl,curl_download)
 importFrom(dbplyr,remote_con)
 importFrom(dplyr,anti_join)
 importFrom(dplyr,arrange)
