### 2019-12-12 254abc0
declare curl dependencies
