+++ R/td_create.R
@@ -38,7 +38,7 @@
 #' @importFrom DBI dbConnect dbDisconnect dbListTables
 #' @importFrom arkdb unark streamable_readr_tsv
 #' @importFrom readr cols
-#' @importFrom curl curl_download
+#' @importFrom curl curl_download curl_fetch_memory
 #' @examples
 #' \donttest{
 #'   \dontshow{
