+++ NAMESPACE
@@ -33,6 +33,7 @@ importFrom(DBI,dbWriteTable)
 importFrom(arkdb,streamable_readr_tsv)
 importFrom(arkdb,unark)
 importFrom(curl,curl_download)
+importFrom(curl,curl_fetch_memory)
 importFrom(dbplyr,remote_con)
 importFrom(dplyr,anti_join)
 importFrom(dplyr,arrange)
