+++ NAMESPACE
@@ -54,6 +54,7 @@ importFrom(dplyr,tbl)
 importFrom(dplyr,top_n)
 importFrom(dplyr,ungroup)
 importFrom(dplyr,union)
+importFrom(jsonlite,fromJSON)
 importFrom(magrittr,"%>%")
 importFrom(memoise,cache_filesystem)
 importFrom(memoise,memoise)
