+++ paper/manuscript.Rmd
@@ -113,18 +113,7 @@ td_create("all")
 
 `taxadb` functions like `get_ids()` and `td_create()` take an optional argument, `db`, to an external database connection.  `taxadb` will work with most DBI-compliant databases such as MySQL or Postgres [see @DBI], but will be much faster when using a column-oriented database engine such as `duckdb` or `MonetDBLite`.  These latter options are also much easier for most users, since each can be installed directly as an R package. `taxadb` will default to the fastest available option.  `taxadb` can also run without a database backend by setting `db=NULL`, though some functions will require a lot (2-20 GB) of free RAM for this to work with many of the large names providers.  
 
-<!--
-`taxadb` does not require the use of an external database system, but will automatically make use of the fastest database server available. `taxadb` can also use any DBI compliant database accessible from R, such as MySQL or Postgres [see @DBI], though the default `duckdb` or `MonetDBLite` will be both significantly faster and easier to use.  
--->
-
-<!--  Next 2 paragraphs should possibly be left to a manual, but are likely to be overlooked aspects there... -->
 By default `taxadb` creates a MonetDBLite database instance [@monetdblite], a columnar-oriented relational database requiring no additional installation while also providing persistent disk-based storage. `taxadb` will automatically handle opening, caching, and closing the database connection. 
-
-<!-- 
-now sidestepped
-
-but note that one limitation of MonetDBLite is the restriction against concurrent access: two separate R sessions on the same machine cannot both access the database at the same time. Users who need concurrent access from multiple sessions may wish to use a standalone MonetDB server or other database server instead.
--->
 
 Functions in `taxadb` are organized into several families: 
 
@@ -193,8 +182,6 @@ This results in many more matches, as different scientific names are recognized 
 
 
 
-
-
 \dummy{\Begin{tcolorbox}[title= Box 1: Taxonomic Identifiers and Synonyms, lower separated=false]}
 
 `get_ids()` returns the `acceptedNameUsageID`, the identifier associated with the *accepted* name.  Some naming providers, such as ITIS and NCBI, provide taxonomic identifiers to both synonyms and accepted names.  Other providers, such as COL and GBIF, only provide identifiers for accepted names.  Common practice in Darwin Core archives is to provide an `acceptedNameUsageID` only for names which are synonyms, and otherwise to provide a `taxonID`.  For accepted names, the `acceptedNameUsageID` is then given as missing (`NA`), while for synonyms, the `taxonID` may be missing (`NA`).  In contrast, `taxadb` lists the `acceptedNameUsageID` for accepted names (where it matches the `taxonID`), as well as known synonyms.  This is semantically identical, but also more convenient for database interfaces, since it allows a name to mapped to it's accepted identifier (or an identifier to map to it's accepted name usage) without the additional logic.  For consistency, we will use the term "identifier" to mean the `acceptedNameUsageID` rather than the more ambiguous `taxonID` (which is undefined for synonyms listed by many providers), unless explicitly stated otherwise.
@@ -203,7 +190,7 @@ This results in many more matches, as different scientific names are recognized 
 
 ## Unresolved names
 
[…]
