+++ R/filter_common.R
@@ -23,6 +23,7 @@ filter_common <- function(name,
                 provider =c("itis", "ncbi", "col", "tpl",
                             "gbif", "fb", "slb", "wd", "ott",
                             "iucn"),
+                version = latest_release(),
                 collect = TRUE,
                 ignore_case = TRUE,
                 db = td_connect()){
@@ -34,6 +35,7 @@ filter_common <- function(name,
             by = "vernacularName",
             provider = provider,
             schema = "common",
+            version = version,
             collect = collect,
             db = db,
             ignore_case = ignore_case)
