+++ R/filter_rank.R
@@ -24,6 +24,7 @@ filter_rank <- function(name,
                     provider = c("itis", "ncbi", "col", "tpl",
                                  "gbif", "fb", "slb", "wd", "ott",
                                  "iucn"),
+                    version = latest_release(),
                     collect = TRUE,
                     ignore_case = TRUE,
                     db = td_connect()){
@@ -31,6 +32,7 @@ filter_rank <- function(name,
   filter_by(x = name,
             by = rank,
             provider = match.arg(provider),
+            version = version,
             collect = collect,
             db = db,
             ignore_case = ignore_case)
