+++ R/filter_name.R
@@ -35,6 +35,7 @@ filter_name <- function(name,
                 provider =c("itis", "ncbi", "col", "tpl",
                             "gbif", "fb", "slb", "wd", "ott",
                             "iucn"),
+                version = latest_release(),
                 collect = TRUE,
                 ignore_case = TRUE,
                 db = td_connect()){
@@ -42,6 +43,7 @@ filter_name <- function(name,
   filter_by(x = name,
             by = "scientificName",
             provider = match.arg(provider),
+            version = version,
             collect = collect,
             db = db,
             ignore_case = ignore_case)
