+++ R/filter_by.R
@@ -11,6 +11,8 @@
 #'  Default is 'itis'.
 #' @param schema One of "dwc" (for Darwin Core data) or "common"
 #' (for the Common names table.)
+#' @param version Which version of the taxadb provider database should we use?
+#'  defaults to latest.  See [available_releases()] for details.
 #' @param collect logical, default `TRUE`. Should we return an in-memory
 #' data.frame (default, usually the most convenient), or a reference to
 #' lazy-eval table on disk (useful for very large tables on which we may
@@ -49,13 +51,14 @@ filter_by <- function(x,
                                    "gbif", "fb", "slb", "wd", "ott",
                                    "iucn"),
                       schema = c("dwc", "common"),
+                      version = latest_release(),
                       collect = TRUE,
                       db = td_connect(),
                       ignore_case = TRUE){
 
   provider <- match.arg(provider)
   schema <- match.arg(schema)
-  db_tbl <- dplyr::mutate(taxa_tbl(provider, schema, db), input = !!sym(by))
+  db_tbl <- dplyr::mutate(taxa_tbl(provider, schema, version, db), input = !!sym(by))
 
   if(ignore_case){
     original <- tibble::tibble(input = x, sort = seq_along(x))
