### 2019-12-18 63fcf16
consistent terminology
version, latest_version, avialable_versions
