+++ R/filter_by.R
@@ -12,7 +12,7 @@
 #' @param schema One of "dwc" (for Darwin Core data) or "common"
 #' (for the Common names table.)
 #' @param version Which version of the taxadb provider database should we use?
-#'  defaults to latest.  See [available_releases()] for details.
+#'  defaults to latest.  See [available_versions()] for details.
 #' @param collect logical, default `TRUE`. Should we return an in-memory
 #' data.frame (default, usually the most convenient), or a reference to
 #' lazy-eval table on disk (useful for very large tables on which we may
@@ -51,7 +51,7 @@ filter_by <- function(x,
                                    "gbif", "fb", "slb", "wd", "ott",
                                    "iucn"),
                       schema = c("dwc", "common"),
-                      version = latest_release(),
+                      version = latest_version(),
                       collect = TRUE,
                       db = td_connect(),
                       ignore_case = TRUE){
