+++ R/filter_rank.R
@@ -24,7 +24,7 @@ filter_rank <- function(name,
                     provider = c("itis", "ncbi", "col", "tpl",
                                  "gbif", "fb", "slb", "wd", "ott",
                                  "iucn"),
-                    version = latest_release(),
+                    version = latest_version(),
                     collect = TRUE,
                     ignore_case = TRUE,
                     db = td_connect()){
