+++ man/available_releases.Rd
@@ -1,14 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/td_create.R
-\name{available_releases}
-\alias{available_releases}
-\title{List available releases}
-\usage{
-available_releases()
-}
-\description{
-List available releases
-}
-\examples{
-available_releases()
-}
