+++ man/synonyms.Rd
@@ -7,6 +7,7 @@
 synonyms(
   name,
   provider = c("itis", "ncbi", "col", "tpl", "gbif", "fb", "slb", "wd", "ott", "iucn"),
+  version = latest_version(),
   collect = TRUE,
   db = td_connect()
 )
@@ -16,6 +17,9 @@ synonyms(
 
 \item{provider}{from which provider should the hierarchy be returned?
 Default is 'itis'.}
+
+\item{version}{Which version of the taxadb provider database should we use?
+defaults to latest.  See \code{\link[=available_versions]{available_versions()}} for details.}
 
 \item{collect}{logical, default \code{TRUE}. Should we return an in-memory
 data.frame (default, usually the most convenient), or a reference to
