+++ man/filter_id.Rd
@@ -8,6 +8,7 @@ filter_id(
   id,
   provider = c("itis", "ncbi", "col", "tpl", "gbif", "fb", "slb", "wd", "ott", "iucn"),
   type = c("taxonID", "acceptedNameUsageID"),
+  version = latest_version(),
   collect = TRUE,
   db = td_connect()
 )
@@ -20,6 +21,9 @@ Default is 'itis'.}
 
 \item{type}{id type.  Can be \code{taxonID} or \code{acceptedNameUsageID},
 see details.}
+
+\item{version}{Which version of the taxadb provider database should we use?
+defaults to latest.  See \code{\link[=available_versions]{available_versions()}} for details.}
 
 \item{collect}{logical, default \code{TRUE}. Should we return an in-memory
 data.frame (default, usually the most convenient), or a reference to
