+++ man/taxa_tbl.Rd
@@ -37,7 +37,8 @@ Return a reference to a given table in the taxadb database
   names <- clean_names(c("Steller's jay", "coopers Hawk"),
                 binomial_only = FALSE, remove_sp = FALSE, remove_punc = TRUE)
 
-  #Get cleaned common names from a provider and search for cleaned names in that table
+  #Get cleaned common names from a provider and
+  # search for cleaned names in that table
   taxa_tbl("itis", "common") \%>\%
   mutate_db(clean_names, "vernacularName", "vernacularNameClean",
             binomial_only = FALSE, remove_sp = FALSE, remove_punc = TRUE) \%>\%
