### 2019-12-18 eabf1c7
json metadata for easier diff
