+++ data-raw/metadata-generator.R
@@ -3,6 +3,8 @@ library(fs)
 library(EML)
 library(tidyverse)
 library(uuid)
+library(here)
+library(jsonlite)
 
 file_hash <- function(x, method = openssl::sha256, ...){
   con <- lapply(x, file, ...)
@@ -10,14 +12,15 @@ file_hash <- function(x, method = openssl::sha256, ...){
   unlist(lapply(hash, as.character))
 }
 
-release <- "2019"
+release <- here::here("2019")
 meta <- fs::dir_info(release) %>%
   mutate(sha256 = map_chr(path, file_hash, raw = TRUE),
          name = fs::path_file(path)) %>%
-  select(path, size, sha256, dateCreated = modification_time)
+  select(name, size, sha256, dateCreated = modification_time, path)
 
 ## Lightweight file metadata
-write_csv(meta, fs::path(release, "meta.csv"))
+write_csv(meta, here::here("data-raw/meta.csv"))
+write_json(tmp, here::here("data-raw/meta.json"), pretty = TRUE, auto_unbox=TRUE)
 
 
 dwc_terms <-
