### 2019-12-18 bad4bc8
minimal json metadata record
