### 2020-01-09 #344 @maelle
@karinorman I've invited you to the rOpensci GitHub organization. Once you've accepted it, you can [transfer the repo](https://help.github.com/en/github/administering-a-repository/transferring-a-repository9 as @ldecicco-USGS said, from the Settings tab of your repo. Once you've transferred the repo, please ping me here, so I can give you admin rights to the repo again. Thank you!
