### 2020-01-09 #344 @stefaniebutland
Congratulations @karinorman! Given that the subject area of taxadb is an important one for rOpenSci and our user/developer community, would you consider writing a blog post?
