### 2020-01-09 #344 @ldecicco-USGS
To echo @stefaniebutland 's request...I think the development of this package would make a good blog story. Others may have different ideas about why `taxadb` is neat, but for me it was how to make a relatively complicated database accessible to users who aren't necessarily subject matter experts. The logistics of creating the local database for the fast connections are really pertinent to other rOpenSci users/developers.
