### 2020-01-10 #344 @karinorman
@ldecicco-USGS Wonderful to hear - thank you so much for making this a smooth process, and thanks again to reviewers @lindsayplatt and @mcsiple for your feedback that greatly improved the package! I'm out of the office until next week but will go through the onboarding steps then. @stefaniebutland I would also be happy to write a blog post, thanks for the invitation!
