### 2020-01-10 #344 @stefaniebutland
Awesome :-) and thank you.
When you're back in the office Kari, please suggest a date by which you'd like to submit a draft post. Typically you submit via pull request, I review the post, and we publish one week after submission.

Instructions are here https://github.com/ropensci/roweb2#contributing-a-blog-post
and I'm happy to answer any questions.
