+++ paper/manuscript.Rmd
@@ -45,9 +45,9 @@ library(knitr)
 ## https://community.rstudio.com/t/why-does-na-match-na-when-joining-two-dataframes/28785/3
 pkgconfig::set_config("dplyr::na_matches" = "never")
 
-printtable <- function(df){
+printtable <- function(df, ...){
   df %>%
-    kableExtra::kable("latex", booktabs=T) %>%
+    kableExtra::kable("latex", booktabs=T, ...) %>%
     kableExtra::kable_styling(full_width = F)
 }
 knitr::opts_chunk$set(cache=TRUE, message=FALSE, warning=FALSE)
@@ -103,7 +103,7 @@ get_ids("Gadus morhua")
 
 This one-time setup will download, uncompress, and import the data into persistant storage (using the appropriate location specified by the operating sytstem [see @rappdirs], or configured using the environmental variable `TAXADB_HOME`).  The example above searches for names in ITIS, the default provider, which can be configured using the `provider` argument. Any future function calls to this function or any other function using data from the same names provider will be able to access this data rapidly without the need for processing or an internet connection.  
 
-Users can also explicitly trigger this one-time setup using `td_create()` and specifying the provider abbreviation (see Table 1), or simply using `all` to install all available providers:
+Users can also explicitly trigger this one-time setup using `td_create()` and specifying the provider abbreviation (see Table 6), or simply using `all` to install all available providers:
 
 
 ```{r eval=FALSE, message=FALSE}
@@ -128,19 +128,19 @@ Functions in `taxadb` are organized into several families:
 Taxonomic identifiers provide a fundamental abstraction which lies at the heart of managing taxonomic names.  For instance, by resolving scientific names to identifiers, we can idetnify which names are synonyms -- different scientific names used to describe the same species -- and which names are not recognized. Unmatched names may indicate an error in data entry or otherwise warrant further investigation. Taxon identifiers are also easily resolved to the original authority (scientific publication) establishing the name. (The historical practice of appending an author and year to a scientific name, e.g. *Poa annua ssp. annua* (Smith 1912), serves a valuable role in disambiguating different uses of the same name but can be notoriously harder to resolve to the appropriate reference, while variation in this convention creates many distinct versions of the same name [@Patterson2010]).  
 
 
-These issues are best illustrated using a minimal example.  We'll consider the task of combining data on bird extinction risk as assessed by the IUCN [@iucn] with data on average adult biomass, as estimated in the Elton Traits v1.0 database [@elton-traits.]  To keep the example concise enough for for visual presentation we will focus on a subset involving just 10 species:
+These issues are best illustrated using a minimal example.  We'll consider the task of combining data on bird extinction risk as assessed by the IUCN [@iucn] with data on average adult biomass, as estimated in the Elton Traits v1.0 database [@elton-traits.]  To keep the example concise enough for for visual presentation we will focus on a subset involving just 10 species [\@ref(tab:iucn_table)]
 
 ```{r message=FALSE}
 trait_data <- read_tsv(system.file("extdata", "trait_data.tsv", package="taxadb"))
 status_data <- read_tsv(system.file("extdata", "status_data.tsv", package="taxadb"))
 ```
 
-```{r echo=FALSE, cache = FALSE}
-status_data %>% printtable()
+```{r iucn_table, echo=FALSE, cache = FALSE}
+status_data %>% printtable(caption = "IUCN status for the selected 10 species.")
 ```
 
-```{r echo = FALSE, cache = FALSE }
-trait_data %>% printtable()
+```{r trait_table, echo = FALSE, cache = FALSE }
+trait_data %>% printtable(caption = "Trait data for the selected 10 species.")
 ```
 
 
[…]
