### 2020-01-21 b643534
Update ignore to get rid of docs
