+++ man/common_starts_with.Rd
@@ -4,13 +4,8 @@
 \alias{common_starts_with}
 \title{common name starts with}
 \usage{
-common_starts_with(
-  name,
-  provider = "itis",
-  version = latest_version(),
-  db = td_connect(),
-  ignore_case = TRUE
-)
+common_starts_with(name, provider = "itis", version = latest_version(),
+  db = td_connect(), ignore_case = TRUE)
 }
 \arguments{
 \item{name}{vector of names (scientific or common, see \code{by})
