+++ man/filter_rank.Rd
@@ -4,15 +4,9 @@
 \alias{filter_rank}
 \title{Get all members (descendants) of a given rank level}
 \usage{
-filter_rank(
-  name,
-  rank,
-  provider = c("itis", "ncbi", "col", "tpl", "gbif", "fb", "slb", "wd", "ott", "iucn"),
-  version = latest_version(),
-  collect = TRUE,
-  ignore_case = TRUE,
-  db = td_connect()
-)
+filter_rank(name, rank, provider = c("itis", "ncbi", "col", "tpl",
+  "gbif", "fb", "slb", "wd", "ott", "iucn"), version = latest_version(),
+  collect = TRUE, ignore_case = TRUE, db = td_connect())
 }
 \arguments{
 \item{name}{taxonomic scientific name (e.g. "Aves")}
@@ -55,10 +49,8 @@ filter_rank("Aves", "class")
 
 }
 \seealso{
-Other filter_by: 
-\code{\link{filter_by}()},
-\code{\link{filter_common}()},
-\code{\link{filter_id}()},
-\code{\link{filter_name}()}
+Other filter_by: \code{\link{filter_by}},
+  \code{\link{filter_common}}, \code{\link{filter_id}},
+  \code{\link{filter_name}}
 }
 \concept{filter_by}
