+++ man/taxa_tbl.Rd
@@ -4,12 +4,9 @@
 \alias{taxa_tbl}
 \title{Return a reference to a given table in the taxadb database}
 \usage{
-taxa_tbl(
-  provider = c("itis", "ncbi", "col", "tpl", "gbif", "fb", "slb", "wd", "ott", "iucn"),
-  schema = c("dwc", "common"),
-  version = latest_version(),
-  db = td_connect()
-)
+taxa_tbl(provider = c("itis", "ncbi", "col", "tpl", "gbif", "fb", "slb",
+  "wd", "ott", "iucn"), schema = c("dwc", "common"),
+  version = latest_version(), db = td_connect())
 }
 \arguments{
 \item{provider}{from which provider should the hierarchy be returned?
