+++ man/td_create.Rd
@@ -4,15 +4,9 @@
 \alias{td_create}
 \title{create a local taxonomic database}
 \usage{
-td_create(
-  provider = "itis",
-  schema = c("dwc", "common"),
-  version = latest_version(),
-  overwrite = FALSE,
-  lines = 1e+05,
-  dbdir = taxadb_dir(),
-  db = td_connect(dbdir)
-)
+td_create(provider = "itis", schema = c("dwc", "common"),
+  version = latest_version(), overwrite = FALSE, lines = 1e+05,
+  dbdir = taxadb_dir(), db = td_connect(dbdir))
 }
 \arguments{
 \item{provider}{a list (character vector) of provider to be included in the
