### 2020-01-21 8a6523c
remove remotes from description
