+++ README.md
@@ -5,11 +5,11 @@
 
 [![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
 [![Travis build
-status](https://travis-ci.org/cboettig/taxadb.svg?branch=master)](https://travis-ci.org/cboettig/taxadb)
+status](https://travis-ci.org/ropensci/taxadb.svg?branch=master)](https://travis-ci.org/ropensci/taxadb)
 [![AppVeyor build
-status](https://ci.appveyor.com/api/projects/status/github/cboettig/taxadb?branch=master&svg=true)](https://ci.appveyor.com/project/cboettig/taxadb)
+status](https://ci.appveyor.com/api/projects/status/github/ropensci/taxadb?branch=master&svg=true)](https://ci.appveyor.com/project/cboettig/taxadb)
 [![Coverage
-status](https://codecov.io/gh/cboettig/taxadb/branch/master/graph/badge.svg)](https://codecov.io/github/cboettig/taxadb?branch=master)
+status](https://codecov.io/gh/ropensci/taxadb/branch/master/graph/badge.svg)](https://codecov.io/github/ropensci/taxadb?branch=master)
 [![CRAN
 status](https://www.r-pkg.org/badges/version/taxadb)](https://cran.r-project.org/package=taxadb)
 
@@ -35,7 +35,7 @@ formats.
 To get started, install the development version directly from GitHub:
 
 ``` r
-devtools::install_github("cboettig/taxadb")
+devtools::install_github("ropensci/taxadb")
 ```
 
 ``` r
@@ -70,16 +70,23 @@ Catalogue of Life:
 birds <- bbs %>% 
   select(species) %>% 
   mutate(id = get_ids(species, "col"))
+#> Warning: `chr_along()` is deprecated as of rlang 0.2.0.
+#> This warning is displayed once per session.
+#> Warning: `overscope_eval_next()` is deprecated as of rlang 0.2.0.
+#> Please use `eval_tidy()` with a data mask instead.
+#> This warning is displayed once per session.
+#> Warning: `overscope_clean()` is deprecated as of rlang 0.2.0.
+#> This warning is displayed once per session.
 
 head(birds, 10)
 #>                          species           id
-#> 1         Dendrocygna autumnalis         <NA>
+#> 1         Dendrocygna autumnalis COL:35517330
 #> 2            Dendrocygna bicolor COL:35517332
 #> 3                Anser canagicus COL:35517329
 #> 4             Anser caerulescens COL:35517325
 #> 5  Chen caerulescens (blue form)         <NA>
 #> 6                   Anser rossii COL:35517328
-#> 7                Anser albifrons         <NA>
+#> 7                Anser albifrons COL:35517308
 #> 8                Branta bernicla COL:35517301
 #> 9      Branta bernicla nigricans COL:35537100
 #> 10             Branta hutchinsii COL:35536445
@@ -89,7 +96,7 @@ Note that some names cannot be resolved to an identifier. This can occur
 because of miss-spellings, non-standard formatting, or the use of a
 synonym not recognized by the naming provider. Names that cannot be
 uniquely resolved because they are known synonyms of multiple different
-species will also return `NA`. The `by_name` filtering functions can
+species will also return `NA`. The `filter_name` filtering functions can
 help us resolve this last case (see below).
 
 `get_ids()` returns the IDs of accepted names, that is
@@ -99,13 +106,13 @@ help us resolve this last case (see below).
 birds %>% 
   mutate(accepted_name = get_names(id, "col")) %>% 
   head()
-#>                         species           id       accepted_name
-#> 1        Dendrocygna autumnalis         <NA>                <NA>
[…]
