+++ README.Rmd
@@ -15,9 +15,9 @@ taxadb:::td_disconnect()
 # taxadb  <img src="man/figures/logo.svg" align="right" alt="" width="120" />
 
 [![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
-[![Travis build status](https://travis-ci.org/cboettig/taxadb.svg?branch=master)](https://travis-ci.org/cboettig/taxadb)
-[![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/cboettig/taxadb?branch=master&svg=true)](https://ci.appveyor.com/project/cboettig/taxadb)
-[![Coverage status](https://codecov.io/gh/cboettig/taxadb/branch/master/graph/badge.svg)](https://codecov.io/github/cboettig/taxadb?branch=master)
+[![Travis build status](https://travis-ci.org/ropensci/taxadb.svg?branch=master)](https://travis-ci.org/ropensci/taxadb)
+[![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/ropensci/taxadb?branch=master&svg=true)](https://ci.appveyor.com/project/cboettig/taxadb)
+[![Coverage status](https://codecov.io/gh/ropensci/taxadb/branch/master/graph/badge.svg)](https://codecov.io/github/ropensci/taxadb?branch=master)
 [![CRAN status](https://www.r-pkg.org/badges/version/taxadb)](https://cran.r-project.org/package=taxadb)
 
 
@@ -31,7 +31,7 @@ Existing approaches to these problems typically rely on web APIs, which can make
 To get started, install the development version directly from GitHub:
 
 ```{r eval=FALSE}
-devtools::install_github("cboettig/taxadb")
+devtools::install_github("ropensci/taxadb")
 ```
 
 
@@ -173,3 +173,5 @@ if(require(codemetar)) codemetar::write_codemeta()
 
 Please note that this project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md).
 By participating in this project you agree to abide by its terms.
+
+[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
