+++ DESCRIPTION
@@ -50,6 +50,6 @@ Suggests:
     MonetDBLite
 Language: en-US
 VignetteBuilder: knitr
-URL: https://docs.ropensci.org/taxadb (website) https://github.com/ropensci/taxadb
+URL: https://docs.ropensci.org/taxadb, https://github.com/ropensci/taxadb
 BugReports: https://github.com/ropensci/taxadb/issues
 AdditionalRepositories: https://github.cboettig.io/drat
