### 2020-01-23 0a64fa5
Update documentation to new roxygen preferance
