+++ man/common_contains.Rd
@@ -4,8 +4,13 @@
 \alias{common_contains}
 \title{common name starts with}
 \usage{
-common_contains(name, provider = "itis", version = latest_version(),
-  db = td_connect(), ignore_case = TRUE)
+common_contains(
+  name,
+  provider = "itis",
+  version = latest_version(),
+  db = td_connect(),
+  ignore_case = TRUE
+)
 }
 \arguments{
 \item{name}{vector of names (scientific or common, see \code{by})
