### 2020-01-23 04f8404
Minor typos/adjustments
