+++ paper/manuscript.Rmd
@@ -60,19 +60,20 @@ options(scipen=999)
 
 
 As ecologists and evolutionary biologists synthesize datasets across larger and larger assemblies of species, we face a continual challenge of reconciling taxonomic names. How many species are in the combined data? Do the studies use the same names for the same species, or do they use different synonyms for the same species?
-Failing to correct for such differences can lead to significant inflation of species counts and miss-aligned  datasets (Figure 1).
-These challenges have become particularly acute as it becomes increasingly common for researchers to work across larger number and diversity of species in any given analysis, which may preclude the resources or substantative taxonomic expertise all clades needed to resolve scientific names [@Patterson2010].
-While these issues have long been recognized in the literature [@boyle2013; @dayrat2005; @bortolus2008, @maldonado2015], and a growing number of databases and tools have emerged and grown over the past few decades [e.g. @itis; @ncbi; @col; @rees2014; @alvarez2018; @wagner2016; @foster2018], it remains difficult to resolve taxonomic names to a common authority in a transparent, efficient, and automatable manner.
+Failing to correct for such differences can lead to significant inflation of species counts and miss-aligned datasets.
+These challenges have become particularly acute as it becomes increasingly common for researchers to work across a larger number and diversity of species in any given analysis, which may preclude the resources or substantative taxonomic expertise for all clades needed to resolve scientific names [@Patterson2010].  
+
+While these issues have long been recognized in the literature [@boyle2013; @dayrat2005; @bortolus2008; @maldonado2015], and a growing number of databases and tools have emerged and grown over the past few decades [e.g. @itis; @ncbi; @col; @rees2014; @alvarez2018; @wagner2016; @foster2018], it remains difficult to resolve taxonomic names to a common authority in a transparent, efficient, and automatable manner.
 Here, we present an R package, `taxadb`, which seeks to address this gap.
 
 Databases of taxonomic names such as the Integrated Taxonomic Information System [ITIS; @itis], the National Center for Biological Information's (NCBI) Taxonomy database, [@ncbi], the Catalogue of Life [COL; @col], and over one hundred other providers have sought to address these problems by providing expert-curated lists of accepted taxonomic names, synonyms, associated taxonomic rank, hierarchical classification, and scientific authority (e.g. author and date) establishing a scientific name.
 The R language [@R] is widely used in ecology and evolution [@Lai2019] and the `taxize` package [@Chamberlain2013] has become a popular way for R users to interact with naming providers and name resolution services.  `taxize` implements bindings to the web APIs (Application Programming Interface) hosted by many popular taxonomic name providers. Unfortunately, this means that functions in the `taxize` are impacted by several major drawbacks that are inherent in the implementation of these central API servers, such as:
 
-- Queries require internet access at all times
-- Queries are slow and inefficient to implement and perform; frequently requiring separate API calls for each taxonomic name
-- The type of query is highly limited by the API design. For instance, it is usually impossible to make queries across the entire corpus of names, such as "which accepted name has the most known synonyms?"
+- Queries require internet access at all times.
+- Queries are slow and inefficient to implement and perform; frequently requiring separate API calls for each taxonomic name.
+- The type of query is highly limited by the API design. For instance, it is usually impossible to make queries across the entire corpus of names, such as "which accepted name has the most known synonyms?".
 - Both query formats and responses differ substantially across different naming providers, making it difficult to apply a script designed for one provider to different provider.
[…]
