### 2020-01-24 9828b44
- update README urls - fast build for vignette
