+++ README.md
@@ -47,8 +47,8 @@ Create a local copy of the Catalogue of Life (2018) database:
 
 ``` r
 td_create("col")
-#> not overwriting dwc_col
-#> not overwriting common_col
+#> not overwriting 2019_dwc_col
+#> not overwriting 2019_common_col
 ```
 
 Read in the species list used by the Breeding Bird Survey:
@@ -70,13 +70,7 @@ Catalogue of Life:
 birds <- bbs %>% 
   select(species) %>% 
   mutate(id = get_ids(species, "col"))
-#> Warning: `chr_along()` is deprecated as of rlang 0.2.0.
-#> This warning is displayed once per session.
-#> Warning: `overscope_eval_next()` is deprecated as of rlang 0.2.0.
-#> Please use `eval_tidy()` with a data mask instead.
-#> This warning is displayed once per session.
-#> Warning: `overscope_clean()` is deprecated as of rlang 0.2.0.
-#> This warning is displayed once per session.
+#> not overwriting 2019_dwc_col
 
 head(birds, 10)
 #>                          species           id
@@ -106,6 +100,7 @@ help us resolve this last case (see below).
 birds %>% 
   mutate(accepted_name = get_names(id, "col")) %>% 
   head()
+#> not overwriting 2019_dwc_col
 #>                         species           id          accepted_name
 #> 1        Dendrocygna autumnalis COL:35517330 Dendrocygna autumnalis
 #> 2           Dendrocygna bicolor COL:35517332    Dendrocygna bicolor
@@ -134,6 +129,7 @@ identifier in ITIS:
 
 ``` r
 get_ids("Trochalopteron henrici gucenense") 
+#> not overwriting 2019_dwc_itis
 #> [1] NA
 ```
 
@@ -142,26 +138,28 @@ zero matches, but to more than one match:
 
 ``` r
 filter_name("Trochalopteron henrici gucenense") 
+#> not overwriting 2019_dwc_itis
 #> # A tibble: 2 x 17
 #>    sort taxonID scientificName taxonRank acceptedNameUsa… taxonomicStatus
 #>   <int> <chr>   <chr>          <chr>     <chr>            <chr>          
 #> 1     1 ITIS:9… Trochaloptero… subspeci… ITIS:916116      synonym        
 #> 2     1 ITIS:9… Trochaloptero… subspeci… ITIS:916117      synonym        
-#> # … with 11 more variables: update_date <chr>, kingdom <chr>,
-#> #   phylum <chr>, class <chr>, order <chr>, family <chr>, genus <chr>,
-#> #   specificEpithet <chr>, vernacularName <chr>,
-#> #   infraspecificEpithet <chr>, input <chr>
+#> # … with 11 more variables: update_date <chr>, kingdom <chr>, phylum <chr>,
+#> #   class <chr>, order <chr>, family <chr>, genus <chr>, specificEpithet <chr>,
+#> #   vernacularName <chr>, infraspecificEpithet <chr>, input <chr>
 ```
 
 ``` r
 filter_name("Trochalopteron henrici gucenense")  %>%
   mutate(acceptedNameUsage = get_names(acceptedNameUsageID)) %>% 
   select(scientificName, taxonomicStatus, acceptedNameUsage, acceptedNameUsageID)
+#> not overwriting 2019_dwc_itis
+#> not overwriting 2019_dwc_itis
 #> # A tibble: 2 x 4
-#>   scientificName       taxonomicStatus acceptedNameUsage   acceptedNameUsa…
-#>   <chr>                <chr>           <chr>               <chr>           
-#> 1 Trochalopteron henr… synonym         Trochalopteron ell… ITIS:916116     
-#> 2 Trochalopteron henr… synonym         Trochalopteron hen… ITIS:916117
[…]
