+++ cran-comments.md
@@ -1,0 +1,14 @@
+We are pleased to submit the intial release of taxadb.  This submission is now less than 5 MB (94 KB).
+
+
+## Test environments
+
+* local OS X install, R 3.6.2
+* ubuntu 16.04 (on travis-ci), R 3.6.2
+* win-builder (devel and release)
+
+## R CMD check results
+
+0 errors | 0 warnings | 0 notes
+
+* This is a new release.
