### 2020-01-24 375fff6
:pencil:
Since when does description mix snake_case
and camelCase?
