### 2020-01-24 f25a90e
fix linking syntax in manual
