+++ README.Rmd
@@ -154,11 +154,11 @@ most_synonyms %>%
 
 ## Learn more
 
-- See richer examples the package [Tutorial](https://docs.ropensci.org/taxadb/articles/intro.html).
+- See richer examples the package [Tutorial](https://docs.ropensci.org/taxadb/articles/articles/intro.html).
 
-- Learn about the underlying data sources and formats in [Data Sources](https://docs.ropensci.org/taxadb/articles/articles/data-sources.html)
+- Learn about the underlying data sources and formats in [Data Sources](https://docs.ropensci.org/taxadb/articles/data-sources.html)
 
-- Get better performance by selecting an alternative [database backend](https://docs.ropensci.org/taxadb/articles/articles/backends.html) engines.
+- Get better performance by selecting an alternative [database backend](https://docs.ropensci.org/taxadb/articles/backends.html) engines.
 
 
 
