+++ cran-comments.md
@@ -1,4 +1,4 @@
-We are pleased to submit the intial release of taxadb.  This submission is now less than 5 MB.
+We are pleased to submit the intial release of taxadb.  This submission is now less than 5 MB. All links should now resolve. Please note that the IUCN Redlist's website sometimes appears to time out on checks.  This is indeed the official website of this major international orgainziation and can easily be confirmed as not a 404 error in any browser.
 
 
 ## Test environments
