+++ DESCRIPTION
@@ -3,7 +3,6 @@ Version: 0.0.2
 Title: A High-Performance Local Taxonomic Database Interface
 Description: Creates a local database of many commonly used taxonomic authorities
              and provides functions that can quickly query this data.
- This package has been peer-reviewed by rOpenSci (v. 0.0.1).
 Authors@R: c(
             person("Carl", "Boettiger", , "cboettig@gmail.com", c("aut", "cre"),
                    comment = c(ORCID = "0000-0002-1642-628X")),
