### 2020-01-30 090cebd
use a local copy of data to test against
