+++ R/td_create.R
@@ -15,22 +15,24 @@
 #' (i.e. updating a local database upon new release.)
 #' @param dbdir a location on your computer where the database
 #' should be installed. Defaults to user data directory given by
-#' [rappdirs::user_data_dir].
+#' `[rappdirs::user_data_dir]`.
 #' @param db connection to a database.  By default, taxadb will set up its own
 #' fast database connection.
 #' @details
-#'  Authorities recognized by taxadb are:
+#'  Authorities currently recognized by taxadb are:
 #'  - `itis`: Integrated Taxonomic Information System, <https://www.itis.gov/>
 #'  - `ncbi`:  National Center for Biotechnology Information,
 #'  <https://www.ncbi.nlm.nih.gov/taxonomy>
 #'  - `col`: Catalogue of Life, <http://www.catalogueoflife.org/>
 #'  - `tpl`: The Plant List, <http://www.theplantlist.org/>
 #'  - `gbif`: Global Biodiversity Information Facility, <https://www.gbif.org/>
-#'  - `fb` FishBase, <http://fishbase.org>
-#'  - `slb`, SeaLifeBase, <http://sealifebase.org>
-#'  - `wd`, Wikidata: <https://www.wikidata.org/>
-#'  - `ott` OpenTree Taxonomy:
+#'  - `fb`: FishBase, <http://fishbase.org>
+#'  - `slb`: SeaLifeBase, <http://sealifebase.org>
+#'  - `wd`: Wikidata: <https://www.wikidata.org/>
+#'  - `ott`: OpenTree Taxonomy:
 #'  <https://github.com/OpenTreeOfLife/reference-taxonomy>
+#'  - `iucn`: IUCN Red List, https://iucnredlist.org
+#'  - `itis_test`: a small subset of ITIS, cached locally with the package for testing purposes only
 #' @return path where database has been installed (invisibly)
 #' @export
 #' @importFrom utils download.file
@@ -52,7 +54,7 @@
 #'   td_create("itis", overwrite = TRUE)
 #'
 #' }
-td_create <- function(provider = "itis",
+td_create <- function(provider = getOption("taxadb_default_provider", "itis"),
                       schema = c("dwc", "common"),
                       version = latest_version(),
                       overwrite = FALSE,
@@ -64,9 +66,7 @@ td_create <- function(provider = "itis",
   if(!dir.exists(dbdir))
     dir.create(dbdir, FALSE, TRUE)
 
-  recognized_provider <- c("itis", "ncbi", "col", "tpl",
-                           "gbif", "fb", "slb", "wd", "ott",
-                           "iucn")
+  recognized_provider <- known_providers
   if (any(provider == "all")) {
     provider <- recognized_provider
   }
@@ -82,6 +82,16 @@ td_create <- function(provider = "itis",
   new_dest <- dest
   new_files <- files
 
+  test_db <- grepl("itis_test", new_files)
+  if(any(test_db)){
+
+    file.copy(from = system.file(file.path("extdata", new_files), package = "taxadb"),
+              to = new_dest[test_db])
+    new_dest <- dest[!test_db]
+    new_files <- files[!test_db]
+  }
+
+  ## can this be removed? should be handled by arkdb
   if (!overwrite) {
     drop <- vapply(dest, file.exists, logical(1))
     new_dest <- dest[!drop]
