+++ R/get_ids.R
@@ -43,7 +43,7 @@
 #' @importFrom dplyr pull
 #' @importFrom tibble column_to_rownames
 get_ids <- function(names,
-                    db = known_providers,
+                    db = getOption("taxadb_default_provider", "itis"),
                     format = c("prefix", "bare", "uri"),
                     version = latest_version(),
                     taxadb_db = td_connect(),
