+++ R/filter_common.R
@@ -20,15 +20,12 @@
 #' }
 #'
 filter_common <- function(name,
-                provider =c("itis", "ncbi", "col", "tpl",
-                            "gbif", "fb", "slb", "wd", "ott",
-                            "iucn"),
+                provider = getOption("taxadb_default_provider", "itis"),
                 version = latest_version(),
                 collect = TRUE,
                 ignore_case = TRUE,
                 db = td_connect()){
 
-  provider <- match.arg(provider)
   if(!assert_has_common(provider)) return(NULL)
 
   filter_by(x = name,
