+++ tests/testthat/setup-test_db.R
@@ -5,6 +5,8 @@ test_db <- file.path(tempdir(), "taxadb")
 #dir.create(test_db, showWarnings = FALSE)
 Sys.setenv(TAXADB_HOME=test_db)
 #Sys.setenv(TAXADB_DRIVER="MonetDBLite")
+#Sys.setenv(TAXADB_DRIVER="RSQLite")
+options(taxadb_default_provider = "itis_test")
 
 ## Consider adding some minimal subset of, say, ITIS data file that
 ## would allow offline testing.  Currently pretty much all tests first

[…]
