### 2020-01-30 f811d85
fix bug introduced in get_names
