### 2020-01-30 5d49707
Patch test data (#70)
* use a local copy of data to test against

* cached data

* fix cache filename

bump version

* db = NULL can only test against online itis version

* fix bug introduced in get_names

* rebuild readme, codemeta
