+++ tests/testthat/test-fast.R
@@ -11,17 +11,16 @@ test_that("setup-free calls to basic functions", {
   skip_if_offline()
 
   df <- taxa_tbl(provider = "itis", db = NULL)
-  chameleons <- filter(df, family == "Chamaeleonidae")
+  sp <- filter(df, family == "Hominidae")
 
   suppressWarnings({
-  df <- filter_rank(name = "Aves", rank = "class", db = NULL)
-  species <- filter_name(df$scientificName, db = NULL)
+  df <- filter_rank(name = "Hominidae", rank = "family", provider = "itis", db = NULL)
+  species <- filter_name("Pan troglodytes", provider = "itis", db = NULL)
   })
 
   expect_is(df, "data.frame")
-  expect_is(species, "data.frame")
-  expect_is(chameleons, "data.frame")
+  expect_is(sp, "data.frame")
   expect_gt(dim(df)[1], 1)
-  expect_gt(dim(chameleons)[1], 1)
+  expect_gt(dim(sp)[1], 1)
 
 })
