+++ R/get_names.R
@@ -19,15 +19,12 @@
 #' @export
 #' @importFrom dplyr pull select collect distinct
 get_names <- function(id,
-                      db = c("itis", "ncbi", "col", "tpl",
-                             "gbif", "fb", "slb", "wd", "ott",
-                             "iucn"),
+                      db = getOption("taxadb_default_provider", "itis"),
                       version = latest_version(),
                       format = c("guess", "prefix", "bare", "uri"),
                       taxadb_db = td_connect()
                      ){
   format <- match.arg(format)
-  db <- match.arg(db)
   n <- length(id)
 
 
