+++ R/filter_rank.R
@@ -21,9 +21,7 @@
 #'
 filter_rank <- function(name,
                     rank,
-                    provider = c("itis", "ncbi", "col", "tpl",
-                                 "gbif", "fb", "slb", "wd", "ott",
-                                 "iucn"),
+                    provider = getOption("taxadb_default_provider", "itis"),
                     version = latest_version(),
                     collect = TRUE,
                     ignore_case = TRUE,
@@ -31,7 +29,7 @@ filter_rank <- function(name,
 
   filter_by(x = name,
             by = rank,
-            provider = match.arg(provider),
+            provider = provider,
             version = version,
             collect = collect,
             db = db,
