+++ R/filter_id.R
@@ -1,6 +1,3 @@
-known_providers <- c("itis", "ncbi", "col", "tpl",
-                     "gbif", "fb", "slb", "wd", "ott",
-                     "iucn")
 
 
 #' Return a taxonomic table matching the requested ids
@@ -39,9 +36,7 @@ known_providers <- c("itis", "ncbi", "col", "tpl",
 #'
 #' }
 filter_id <- function(id,
-                  provider = c("itis", "ncbi", "col", "tpl",
-                               "gbif", "fb", "slb", "wd", "ott",
-                               "iucn"),
+                  provider = getOption("taxadb_default_provider", "itis"),
                   type = c("taxonID", "acceptedNameUsageID"),
                   version = latest_version(),
                   collect = TRUE,
@@ -49,7 +44,7 @@ filter_id <- function(id,
 
     filter_by(x = id,
               by = match.arg(type),
-              provider = match.arg(provider),
+              provider = provider,
               version = version,
               collect = collect,
               db = db,
