+++ vignettes/data-sources.Rmd
@@ -30,7 +30,7 @@ knitr::opts_chunk$set(
 `fb`   | [FishBase](https://fishbase.ca)
 `slb`  | [SeaLifeBase](https://www.sealifebase.ca/)
 `wd`   | [WikiData](https://wikidata.org)
-`iucn` | [The IUCN Red List](https://www.iucnredlist.org) of endangered species status
+`iucn` | The IUCN Red List of endangered species status, https://www.iucnredlist.org
 `ott`  | [Open Tree of Life taxonomy](https://tree.opentreeoflife.org/about/taxonomy-version/ott3.1).  
 
 

[…]
