### 2020-01-30 13c57ec
depricate synonyms function
