+++ NAMESPACE
@@ -15,7 +15,6 @@ export(get_names)
 export(mutate_db)
 export(name_contains)
 export(name_starts_with)
-export(synonyms)
 export(taxa_tbl)
 export(td_connect)
 export(td_create)
@@ -41,7 +40,6 @@ importFrom(dplyr,collect)
 importFrom(dplyr,count)
 importFrom(dplyr,distinct)
 importFrom(dplyr,filter)
-importFrom(dplyr,full_join)
 importFrom(dplyr,group_by)
 importFrom(dplyr,left_join)
 importFrom(dplyr,mutate)
