### 2020-01-30 ac4c9b1
Add optional support for duckdb (#71)
* duckdb okay except for call to synonyms()

Would be good to investigate, but is it time to deprecate synonyms() anyway?

* depricate synonyms function
