+++ man/synonyms.Rd
@@ -1,48 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/synonyms.R
-\name{synonyms}
-\alias{synonyms}
-\title{synonyms}
-\usage{
-synonyms(
-  name,
-  provider = getOption("taxadb_default_provider", "itis"),
-  version = latest_version(),
-  collect = TRUE,
-  db = td_connect()
-)
-}
-\arguments{
-\item{name}{a character vector of scientific names, e.g. "Homo sapiens"}
-
-\item{provider}{from which provider should the hierarchy be returned?
-Default is 'itis', which can also be configured using \verb{options(default_taxadb_provider=...")}.
-See \verb{[td_create]} for a list of recognized providers.}
-
-\item{version}{Which version of the taxadb provider database should we use?
-defaults to latest.  See \link{available_versions} for details.}
-
-\item{collect}{logical, default \code{TRUE}. Should we return an in-memory
-data.frame (default, usually the most convenient), or a reference to
-lazy-eval table on disk (useful for very large tables on which we may
-first perform subsequent filtering operations.)}
-
-\item{db}{a connection to the taxadb database. See details.}
-}
-\description{
-Resolve provided list of names against all known synonyms
-}
-\examples{
-\donttest{
-  \dontshow{
-   ## All examples use a temporary directory
-   Sys.setenv(TAXADB_HOME=tempdir())
-  }
-
-sp <- c("Trochalopteron henrici gucenense",
-        "Trochalopteron elliotii")
-synonyms(sp)
-
-}
-
-}
