### 2020-01-31 fb7ee6a
patch way duckdb calls disk storage location
duckdb was attempting in-memory location, since
`duckdb::duckdb()` does not seem to follow the `dbname` convention of MonetDBLite, RSQLite and others in passing the path for persistent storage.  cc @noamross
