+++ R/td_connect.R
@@ -66,7 +66,7 @@ db_driver <- function(dbname, driver = Sys.getenv("TAXADB_DRIVER")){
     MonetDBLite <- getExportedValue("MonetDBLite", "MonetDBLite")
     drivers <- c("MonetDBLite", drivers)
   }
-  ## duckdb lacks necessary capabilities (e.g. temp tables) at this time
+  ## duckdb lacks necessary stability
   ## https://github.com/cwida/duckdb/issues/58
   if (requireNamespace("duckdb", quietly = TRUE)){
     duckdb <- getExportedValue("duckdb", "duckdb")
@@ -78,7 +78,7 @@ db_driver <- function(dbname, driver = Sys.getenv("TAXADB_DRIVER")){
 
 
   db <- switch(driver,
-         duckdb = DBI::dbConnect(duckdb(),
+         duckdb = DBI::dbConnect(duckdb( dbdir = file.path(dbname,"duckdb")),
                                  dbname = file.path(dbname,"duckdb")),
          MonetDBLite = monetdblite_connect(file.path(dbname,"MonetDBLite")),
          RSQLite = DBI::dbConnect(SQLite(),
