+++ README.md
@@ -97,13 +97,13 @@ help us resolve this last case (see below).
 birds %>% 
   mutate(accepted_name = get_names(id, "col")) %>% 
   head()
-#>                         species           id          accepted_name
-#> 1        Dendrocygna autumnalis COL:35517330 Dendrocygna autumnalis
-#> 2           Dendrocygna bicolor COL:35517332    Dendrocygna bicolor
-#> 3               Anser canagicus COL:35517329          Chen canagica
-#> 4            Anser caerulescens COL:35517325      Chen caerulescens
-#> 5 Chen caerulescens (blue form)         <NA>                   <NA>
-#> 6                  Anser rossii COL:35517328            Chen rossii
+#>                         species           id        accepted_name
+#> 1        Dendrocygna autumnalis COL:35517330      Tringa flavipes
+#> 2           Dendrocygna bicolor COL:35517332    Picoides dorsalis
+#> 3               Anser canagicus COL:35517329   Setophaga castanea
+#> 4            Anser caerulescens COL:35517325  Bombycilla cedrorum
+#> 5 Chen caerulescens (blue form)         <NA>       Icteria virens
+#> 6                  Anser rossii COL:35517328 Somateria mollissima
 ```
 
 This illustrates that some of our names, e.g. *Dendrocygna bicolor* are
@@ -126,7 +126,7 @@ identifier in ITIS:
 ``` r
 get_ids("Trochalopteron henrici gucenense") 
 #> Importing 2019_dwc_itis.tsv.bz2 in 100000 line chunks:
-#>  ...Done! (in 55.84305 secs)
+#>  ...Done! (in 13.33816 secs)
 #> [1] NA
 ```
 
@@ -138,8 +138,8 @@ filter_name("Trochalopteron henrici gucenense")
 #> # A tibble: 2 x 17
 #>    sort taxonID scientificName taxonRank acceptedNameUsa… taxonomicStatus
 #>   <int> <chr>   <chr>          <chr>     <chr>            <chr>          
-#> 1     1 ITIS:9… Trochaloptero… subspeci… ITIS:916116      synonym        
-#> 2     1 ITIS:9… Trochaloptero… subspeci… ITIS:916117      synonym        
+#> 1     1 ITIS:9… Trochaloptero… subspeci… ITIS:916117      synonym        
+#> 2     1 ITIS:9… Trochaloptero… subspeci… ITIS:916116      synonym        
 #> # … with 11 more variables: update_date <chr>, kingdom <chr>, phylum <chr>,
 #> #   class <chr>, order <chr>, family <chr>, genus <chr>, specificEpithet <chr>,
 #> #   vernacularName <chr>, infraspecificEpithet <chr>, input <chr>
@@ -152,8 +152,8 @@ filter_name("Trochalopteron henrici gucenense")  %>%
 #> # A tibble: 2 x 4
 #>   scientificName          taxonomicStatus acceptedNameUsage    acceptedNameUsag…
 #>   <chr>                   <chr>           <chr>                <chr>            
-#> 1 Trochalopteron henrici… synonym         Trochalopteron elli… ITIS:916116      
-#> 2 Trochalopteron henrici… synonym         Trochalopteron henr… ITIS:916117
+#> 1 Trochalopteron henrici… synonym         Trochalopteron henr… ITIS:916117      
+#> 2 Trochalopteron henrici… synonym         Trochalopteron elli… ITIS:916116
 ```
 
 Similar functions `filter_id`, `filter_rank`, and `filter_common` take
@@ -165,16 +165,16 @@ filter_rank(name = "Aves", rank = "class", provider = "col")
[…]
