+++ paper/manuscript.Rmd
@@ -261,7 +261,9 @@ tibble(provider = c("Integrated Taxonomic Information System (ITIS 2019)",
 
 ```
 
+## Data Versioning
 
+Some taxonomic name providers (e.g. OTT, COL) offer periodic releases of a static names list, while many other providers (e.g. ITIS, NCBI, FB, IUCN) offer name data on a rolling basis (i.e. the data returned by a given download URL is updated continuously or at arbitrary intervals without any additional indication if and how that data has changed.)  `taxadb`'s `td_create()` function downloads and stores cached snapshopts from each provider, which follow an annual release model to support reproducible analyses.  By default, `taxadb` will determine the latest version available (at the time of writing this is version `2019`), and install and use this version of all snapshots for each provider.  However, all taxadb functions can take   
 
 ## Discussion
 
