### 2020-01-31 a69ab05
bit on versioning added to paper and vignette :pen:
