### 2020-02-03 #344 @stefaniebutland
Hi @karinorman. Will you be able to submit a draft blog post this week?
Happy to answer any questions.
