### 2020-02-06 4618b10
use README as intro vignette, it's better
