+++ vignettes/articles/intro.Rmd
@@ -1,7 +1,7 @@
 ---
 title: "Tutorial for taxadb"
 author: "Carl Boettiger, Kari Norman"
-date: "2020-01-24"
+date: "2020-02-06"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{intro}
@@ -11,385 +11,224 @@ vignette: >
 
 
 
-Here we illustrate the use of `taxadb` to resolve names in the Breeding Bird Survey species list more completely.  To get started, we load the necessary libraries and create the local copy of the `taxadb` database for all naming providers, if we have not already done so:
+
+The goal of `taxadb` is to provide *fast*, *consistent* access to taxonomic data, supporting common tasks such as resolving taxonomic names to identifiers, looking up higher classification ranks of given species, or returning a list of all species below a given rank. These tasks are particularly common when synthesizing data across large species assemblies, such as combining occurrence records with trait records. 
+
+Existing approaches to these problems typically rely on web APIs, which can make them impractical for work with large numbers of species or in more complex pipelines.  Queries and returned formats also differ across the different taxonomic authorities, making tasks that query multiple authorities particularly complex. `taxadb` creates a *local* database of most readily available taxonomic authorities, each of which is transformed into consistent, standard, and researcher-friendly tabular formats.  
+
+
+## Install and initial setup
+
+To get started, install the development version directly from GitHub:
 
 
 ```r
-library(dplyr)
-library(readr)
+devtools::install_github("ropensci/taxadb")
+```
+
+
+
+```r
 library(taxadb)
+library(dplyr) # Used to illustrate how a typical workflow combines nicely with `dplyr`
 ```
+
+Create a local copy of the Catalogue of Life (2018) database: 
 
 
 ```r
-td_create("all")
+td_create("col")
+#> Importing 2019_common_col.tsv.bz2 in 100000 line chunks:
+#> [-] chunk 2[\] chunk 3[|] chunk 4[/] chunk 5	...Done! (in 9.042107 secs)
 ```
 
-Note that by default, `td_create()` will not attempt to overwrite any providers which we have already installed.  This means we can freely include `td_create()` at the beginning of scripts and analysis for reproducibility without concern that it will keep re-downloading and slowing things down.  With this in place, we are ready to load in the species list from the Breeding Bird Survey [@bbs], taken from <ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/SpeciesList.txt>:
 
-
+Read in the species list used by the Breeding Bird Survey:
 
 
 ```r
-bbs <- read_tsv(system.file("extdata/bbs.tsv", package="taxadb"))
-bbs
-#> # A tibble: 750 x 9
-#>      Seq AOU   English_Common_Name  French_Common_Name      Spanish_Common_Name   order    family genus  species       
-#>    <dbl> <chr> <chr>                <chr>                   <chr>                 <chr>    <chr>  <chr>  <chr>         
-#>  1     6 01770 Black-bellied Whist… "Dendrocygne \xe0 vent… Dendrocygna autumnal… Anserif… Anati… Dendr… Dendrocygna a…
-#>  2     7 01780 Fulvous Whistling-D… "Dendrocygne fauve"     Dendrocygna bicolor   Anserif… Anati… Dendr… Dendrocygna b…
[…]
