### 2020-02-06 12e1e97
streamline final example in README
