+++ paper/refs.bib
@@ -369,3 +369,24 @@ and Logs},
 	pages = {973--984},
 	file = {Maldonado et al_2015_Global Ecology and Biogeography.pdf:/Users/karinorman/Dropbox/Literature/Maldonado et al_2015_Global Ecology and Biogeography.pdf:application/pdf}
 }
+
+
+@article{ropensci,
+	title = {Building {Software}, {Building} {Community}: {Lessons} from the {rOpenSci} {Project}},
+	volume = {3},
+	copyright = {Authors who publish with this journal agree to the following terms:    Authors retain copyright and grant the journal right of first publication with the work simultaneously licensed under a  Creative Commons Attribution License  that allows others to share the work with an acknowledgement of the work's authorship and initial publication in this journal.  Authors are able to enter into separate, additional contractual arrangements for the non-exclusive distribution of the journal's published version of the work (e.g., post it to an institutional repository or publish it in a book), with an acknowledgement of its initial publication in this journal.  Authors are permitted and encouraged to post their work online (e.g., in institutional repositories or on their website) prior to and during the submission process, as it can lead to productive exchanges, as well as earlier and greater citation of published work (See  The Effect of Open Access ).  All third-party images reproduced on this journal are shared under Educational Fair Use. For more information on  Educational Fair Use , please see  this useful checklist prepared by Columbia University Libraries .   All copyright  of third-party content posted here for research purposes belongs to its original owners.  Unless otherwise stated all references to characters and comic art presented on this journal are ©, ® or ™ of their respective owners. No challenge to any owner’s rights is intended or should be inferred.},
+	issn = {2049-9647},
+	shorttitle = {Building {Software}, {Building} {Community}},
+	url = {http://openresearchsoftware.metajnl.com/articles/10.5334/jors.bu/},
+	doi = {10.5334/jors.bu},
+	abstract = {rOpenSci is a developer collective originally formed in 2011 by graduate students and post-docs from ecology and evolutionary biology to collaborate on building software tools to facilitate a more open and synthetic approach in the face of transformative rise of large and heterogeneous data. Born on the internet (the collective only began through chance discussions over social media), we have grown into a widely recognized effort that supports an ecosystem of some 45 software packages, engages scores of collaborators, has taught dozens of workshops around the world, and has secured over \$480,000 in grant support. As young scientists working in an academic context largely without direct support for our efforts, we have first hand experience with most of the the technical and social challenges WSSSPE seeks to address. In this paper we provide an experience report which describes our approach and success in building an effective and diverse community.},
+	language = {en},
+	number = {1},
+	urldate = {2020-02-06},
+	journal = {Journal of Open Research Software},
+	author = {Boettiger, Carl and Chamberlain, Scott and Hart, Edmund and Ram, Karthik},
+	month = nov,
+	year = {2015},
+	keywords = {data science, open science, R, ropensci},
+	pages = {e8}
+}
