+++ paper/manuscript.Rmd
@@ -3,8 +3,6 @@ title: "taxadb: A High-Performance Local Taxonomic Database Interface"
 author:
   - name: "Kari Norman"
     affiliation: ucb
-#  - name: "Jorrit Poelen"
-#    affiliation: jorritt
   - name: "Scott Chamberlain"
     affiliation: ropensci      
   - name: "Carl Boettiger"
@@ -12,16 +10,14 @@ author:
 address:
   - code: ucb
     address: "Dept of Environmental Science, Policy, and Management, University of California Berkeley, Berkeley CA 94720-3114, USA"
-# - code: jorrit
-#   address: "Independent consultant, Oakland, CA, USA"
   - code: ropensci
     address: "The rOpenSci Project, University of California Berkeley, Berkeley CA 94720-3114, USA"
 abstract: |
   A familiar and growing challenge in ecological and evolutionary research is that of reconciling scientific names of relevant taxa when combining data from separate sources. While this problem is already well understood and numerous naming authorities have been created to address the issue, most researchers lack a fast, consistent, and intuitive way to reconcile taxonomic names.  Here, we present `taxadb` R package to address this gap.  In contrast to existing tools, `taxadb` provides the following:  
-  1) `taxadb` accesses established naming authorities to resolve synonyms, IDs, and hiearchical classification.  
+  1) `taxadb` accesses established naming authorities to resolve synonyms, IDs, and hierarchical classification.  
   2) `taxadb` creates a local database, managed automatically from within R, that provides fast operations on millions of taxonomic names.  
   3) `taxadb` provides a consistent and intuitive data format.  
-  4) `taxadb` is built on a simple, extensible and language agnostic design that can easily accomodiate new authorities.  
+  4) `taxadb` is built on a simple, extensible and language agnostic design that can easily accommodate new authorities.  
   
 journal: "Methods in Ecology & Evolution"
 date: "`r Sys.Date()`"
@@ -46,29 +42,50 @@ library(knitr)
 ## https://community.rstudio.com/t/why-does-na-match-na-when-joining-two-dataframes/28785/3
 pkgconfig::set_config("dplyr::na_matches" = "never")
 
+## Display NAs as "-" because "NA" looks like a category
+options(knitr.kable.NA = '-') 
+
 printtable <- function(df, ...){
   df %>%
     kableExtra::kable("latex", booktabs=T, ...) %>%
     kableExtra::kable_styling(full_width = F)
 }
-knitr::opts_chunk$set(cache=TRUE, message=FALSE, warning=FALSE)
+knitr::opts_chunk$set(cache=FALSE, message=FALSE, warning=FALSE)
 
 # no sci notation integers pleases
 options(scipen=999)
 
+taxadb::td_disconnect()
 ```
 
 
 
-As ecologists and evolutionary biologists synthesize datasets across larger and larger assemblies of species, we face a continual challenge of reconciling taxonomic names. How many species are in the combined data? Do the studies use the same names for the same species, or do they use different synonyms for the same species?
+As ecologists and evolutionary biologists synthesize datasets across larger and larger assemblies of species,
+we face a continual challenge of reconciling taxonomic names. How many species are in the combined data?
+Do the studies use the same names for the same species, or do they use different synonyms for the same species?
 Failing to correct for such differences can lead to significant inflation of species counts and miss-aligned datasets.
-These challenges have become particularly acute as it becomes increasingly common for researchers to work across a larger number and diversity of species in any given analysis, which may preclude the resources or substantative taxonomic expertise for all clades needed to resolve scientific names [@Patterson2010].  
+These challenges have become particularly acute as it becomes increasingly common for researchers to work
+across a larger number and diversity of species in any given analysis, 
[…]
