### 2020-02-07 c7ecf38
attempt 2 to explore metadata options
