### 2020-02-08 91b34df
Change IUCN example data column name to consistently refer to IUCN by name rather than by "redlist".
