+++ paper/manuscript.Rmd
@@ -48,7 +48,7 @@ options(knitr.kable.NA = '-')
 printtable <- function(df, ...){
   df %>%
     kableExtra::kable("latex", booktabs=T, ...) %>%
-    kableExtra::kable_styling(full_width = F)
+    kableExtra::kable_styling(full_width = F, latex_options = "hold_position")
 }
 knitr::opts_chunk$set(cache=FALSE, message=FALSE, warning=FALSE)
 
@@ -70,7 +70,7 @@ which may preclude the resources or substantive taxonomic expertise for all clad
 needed to resolve scientific names [@Patterson2010].  
 
 While these issues have long been recognized in the literature [@boyle2013; @dayrat2005; @bortolus2008; @maldonado2015],
-and a growing number of databases and tools have emerged and grown over the past few decades
+and a growing number of databases and tools have emerged over the past few decades
 [e.g. @itis; @ncbi; @col; @rees2014; @alvarez2018; @wagner2016; @foster2018],
 it remains difficult to resolve taxonomic names to a common authority in a transparent, efficient, and automatable manner.
 Here, we present an R package, `taxadb`, which seeks to address this gap.
@@ -94,9 +94,9 @@ that are inherent in the implementation of these central API servers, such as:
 - Most queries are not reproducible, as the results depend on the state of the central server (and potentially the quality of the internet connection).  Many names providers update the server data either continuously or at regular intervals, including both revising existing names and adding new names.
 
 Instead of binding existing web APIs, `taxadb` is built around a set of compressed text files which are 
-automatically downloaded and imported and stored on a local database by `taxadb`.  
+automatically downloaded, imported, and stored on a local database by `taxadb`. 
 The largest of the taxonomic naming providers today contain under 6 million name records with uncompressed file sizes
-over a GB, which can be compressed to around 50 MB and downloaded in under a minute on a 1 MB/s connection.  
+over a GB, which can be compressed to around 50 MB and downloaded in under a minute on a 1 MB/s connection. 
 By using a local database as the backend, `taxadb` allows R users to interact with large data files without large memory (RAM) requirements.  A query for a single name over the web API requires a remote server to respond, execute the query, 
 and serialize the response, which can take several seconds. Thus it does not take many taxa before transferring the 
 entire data set to query locally is more efficient.  Moreover, this local copy can be cached on the user's machine, 
@@ -130,8 +130,8 @@ tibble(provider = c("Integrated Taxonomic Information System (ITIS 2019)",
                     description = desc
 ) %>%
   knitr::kable(format = "latex", col.names = c("Provider", "Abbreviation", "Number of \nIdentifiers", "Description"), escape = FALSE, 
-               caption = "Descriptions of providers supported by taxadb.") %>%
-  kable_styling(latex_options="scale_down") %>%
+               caption = "Descriptions of the providers supported by taxadb with their reference abbreviation and the total number of identifiers contained by each provider.") %>%
+  kable_styling(latex_options= c("scale_down", "hold_position")) %>%
   kableExtra::column_spec(1, width = "5cm") %>%
   kableExtra::column_spec(4, width = "10cm")
 
@@ -157,7 +157,7 @@ library(tidyverse)
 library(taxadb)
 ```
 
-After loading the package, we look up the taxonomic identifier for a Atlantic Cod, *Gadus morhua*, and the compliment:
+After loading the package, we look up the taxonomic identifier for Atlantic Cod, *Gadus morhua*, and the compliment:
 
 
 ```{r}
@@ -166,9 +166,9 @@ get_names("ITIS:164712")
 ```
 
 
[…]
