### 2020-02-08 42e6f1c
use the compressed uri as the hash_uri, since that is what the hash-archive.org will use
