+++ data-raw/metadata-generator.R
@@ -13,12 +13,13 @@ file_hash <- function(x, method = openssl::sha256, ...){
 release <- here::here("2019")
 ## slower if we compute hash_uris on uncompressed content!
 meta <- fs::dir_info(release) %>%
-  mutate(sha256_compressed = map_chr(path, file_hash, raw = TRUE),
-         hash_uri = paste0("hash://sha256/", map_chr(path, file_hash)),
+  mutate(sha256_uncompressed = map_chr(path, file_hash, raw = FALSE),
+         sha256_compressed = map_chr(path, file_hash, raw=TRUE),
+         hash_uri = paste0("hash://sha256/", sha256_compressed),
          name = fs::path_file(path),
          contentType = "text/tab-separated-values",
          contentEncoding = "bz2") %>%
-  select(name, size, sha256_compressed, hash_uri,
+  select(name, size, sha256_uncompressed, sha256_compressed, hash_uri,
          dateCreated = modification_time,
          contentType, contentEncoding, path)
 
@@ -27,7 +28,7 @@ meta <- fs::dir_info(release) %>%
 #write_csv(meta, here::here("data-raw/meta.csv"))
 meta %>%
   select(-path) %>%
-  mutate(size = trimws(as.character(size))) %>%
+  #mutate(size = trimws(as.character(size))) %>%
   write_json(here::here("data-raw/meta.json"),
              pretty = TRUE, auto_unbox=TRUE)
 
@@ -89,9 +90,9 @@ dataset = eml$dataset(
   intellectualRights = "",
   abstract =  "",
   dataTable = dataTable,
-  keywordSet = keywordSet,
-  coverage = coverage,
-  methods = methods
+  #keywordSet = keywordSet,
+ # coverage = coverage,
+ # methods = methods
 )
 
 
