+++ paper/coverletterMEE.Rmd
@@ -1,0 +1,19 @@
+---
+author: Kari Norman
+return-address: 
+  - Dept of Environmental Science Policy and Management 
+  - \linebreak University of California Berkeley
+  - \linebreak 130 Hilgard Way
+  - Berkeley, CA 94709
+
+opening: Dear Editors,
+closing: Thank you for your consideration,
+signature: Kari Norman, Scott Chamberlain, Carl Boettiger
+return-email: kari.norman@berkeley.edu
+
+output: komaletter::komaletter
+---
+
+We are submitting our manuscript titled “taxadb: A High-Performance Local Taxonomic Database Interface” to be considered for publication as an Application in Methods in Ecology and Evolution. 
+
+Tracking and reconciling changes in taxonomy across time and between datasets is a core but often overlooked challenge in Ecology and Evolutionary biology. As combining larger and larger datasets becomes increasingly common, the potential for serious species mismatches has important implications for the scientific conclusions drawn. While multiple efforts to standardize taxonomic data exist and and some tools to access those providers have been developed, it remains difficult to resolve names to a taxonomic authority in quick, reproducible way. Here we present a new R package for accessing  taxonomic data using infrastructure that enables easy incorporation of taxonomic data to existing workflows and further exploration of the taxonomic providers. This package was reviewed and accepted by rOpenSci and can be found at www.github.com/ropensci/taxadb. 
