### 2020-02-12 20a4b8d
add acknowledgments and author contributions
