+++ paper/manuscript.Rmd
@@ -1,7 +1,7 @@
 ---
 title: "taxadb: A High-Performance Local Taxonomic Database Interface"
 author:
-  - name: "Kari Norman"
+  - name: "Kari E A Norman"
     affiliation: ucb
   - name: "Scott Chamberlain"
     affiliation: ropensci
@@ -314,6 +314,11 @@ Taxonomic identifiers are a powerful tool for maintaining taxonomic consistency,
 
 Lastly, we note that local database design used in `taxadb` is not unique to taxonomic names.  Despite the rapid expansion of REST API-based interfaces to ecological data [@ropensci], in our experience, much of the data relevant to ecologists and evolutionary biologists today would be also be ameniable to the local database design.  The local database approach is much easier for data providers (who can leverage static scientific database repositories instead of maintaining REST servers) and often much faster for data consumers. 
 
+# Acknowledgments
+We thank the many researchers who contributed to the data and infrastructure of the various taxonomic providers we access through our package. Support for the development of this package was provided by United States Department of Energy through the Computational Sciences Graduate Fellowship (DOE CSGF) under grant number DE-FG02-97ER25308 awarded to K.E.A.N..
+
+# Authors' Contributions
+K.E.A.N., S.C., and C.B. contributed to conceptual development of the package. K.E.A.N. and C.B. developed the package and contributed to the manuscript. 
 
 \pagebreak
 
