+++ README.md
@@ -12,6 +12,7 @@ status](https://ci.appveyor.com/api/projects/status/github/ropensci/taxadb?branc
 status](https://codecov.io/gh/ropensci/taxadb/branch/master/graph/badge.svg)](https://codecov.io/github/ropensci/taxadb?branch=master)
 [![CRAN
 status](https://www.r-pkg.org/badges/version/taxadb)](https://cran.r-project.org/package=taxadb)
+[![peer-review](https://badges.ropensci.org/344_status.svg)](https://github.com/ropensci/software-review/issues/344)
 
 The goal of `taxadb` is to provide *fast*, *consistent* access to
 taxonomic data, supporting common tasks such as resolving taxonomic
@@ -97,13 +98,13 @@ help us resolve this last case (see below).
 birds %>% 
   mutate(accepted_name = get_names(id, "col")) %>% 
   head()
-#>                         species           id        accepted_name
-#> 1        Dendrocygna autumnalis COL:35517330      Tringa flavipes
-#> 2           Dendrocygna bicolor COL:35517332    Picoides dorsalis
-#> 3               Anser canagicus COL:35517329   Setophaga castanea
-#> 4            Anser caerulescens COL:35517325  Bombycilla cedrorum
-#> 5 Chen caerulescens (blue form)         <NA>       Icteria virens
-#> 6                  Anser rossii COL:35517328 Somateria mollissima
+#>                         species           id          accepted_name
+#> 1        Dendrocygna autumnalis COL:35517330 Dendrocygna autumnalis
+#> 2           Dendrocygna bicolor COL:35517332    Dendrocygna bicolor
+#> 3               Anser canagicus COL:35517329          Chen canagica
+#> 4            Anser caerulescens COL:35517325      Chen caerulescens
+#> 5 Chen caerulescens (blue form)         <NA>                   <NA>
+#> 6                  Anser rossii COL:35517328            Chen rossii
 ```
 
 This illustrates that some of our names, e.g. *Dendrocygna bicolor* are
@@ -136,8 +137,8 @@ filter_name("Trochalopteron henrici gucenense")
 #> # A tibble: 2 x 17
 #>    sort taxonID scientificName taxonRank acceptedNameUsa… taxonomicStatus
 #>   <int> <chr>   <chr>          <chr>     <chr>            <chr>          
-#> 1     1 ITIS:9… Trochaloptero… subspeci… ITIS:916117      synonym        
-#> 2     1 ITIS:9… Trochaloptero… subspeci… ITIS:916116      synonym        
+#> 1     1 ITIS:9… Trochaloptero… subspeci… ITIS:916116      synonym        
+#> 2     1 ITIS:9… Trochaloptero… subspeci… ITIS:916117      synonym        
 #> # … with 11 more variables: update_date <chr>, kingdom <chr>, phylum <chr>,
 #> #   class <chr>, order <chr>, family <chr>, genus <chr>, specificEpithet <chr>,
 #> #   vernacularName <chr>, infraspecificEpithet <chr>, input <chr>
@@ -150,8 +151,8 @@ filter_name("Trochalopteron henrici gucenense")  %>%
 #> # A tibble: 2 x 4
 #>   scientificName          taxonomicStatus acceptedNameUsage    acceptedNameUsag…
 #>   <chr>                   <chr>           <chr>                <chr>            
-#> 1 Trochalopteron henrici… synonym         Trochalopteron henr… ITIS:916117      
-#> 2 Trochalopteron henrici… synonym         Trochalopteron elli… ITIS:916116
[…]
