+++ data-raw/col.R
@@ -1,20 +1,17 @@
-library(dplyr)
-library(readr)
-library(forcats)
-library(stringi)
-source("data-raw/helper-routines.R")
 
-# All snapshots available from: http://www.catalogueoflife.org/DCA_Export/archive.php
-
+#' @export
+#' @import stringi forcats readr dplyr
+#' @details NOTE: A list of  all snapshots available from: http://www.catalogueoflife.org/DCA_Export/archive.php
 preprocess_col <- function(url = paste0("http://www.catalogueoflife.org/DCA_Export/zip-fixed/",
                                         2019,
                                         "-annual.zip"),
                            output_paths = c(dwc = "2019/dwc_col.tsv.bz2",
-                                            common = "2019/common_col.tsv.bz2")){
+                                            common = "2019/common_col.tsv.bz2"),
+                           dir = file.path(tempdir(), "col")){
 
-  dir = file.path(tempdir(), "col")
+ 
   dir.create(dir, FALSE, FALSE)
-  download.file(url,
+  curl::curl_download(url,
                 file.path(dir, "col-annual.zip"))
   unzip(file.path(dir, "col-annual.zip"), exdir=dir)
 
@@ -32,7 +29,6 @@ preprocess_col <- function(url = paste0("http://www.catalogueoflife.org/DCA_Expo
     select(taxonID, scientificName, acceptedNameUsageID, taxonomicStatus, taxonRank,
            kingdom, phylum, class, order, family, genus, specificEpithet, infraspecificEpithet,
            taxonConceptID, isExtinct, nameAccordingTo, namePublishedIn, scientificNameAuthorship)
-
   taxa <- bind_rows(
     taxa_tmp %>%
       filter(!is.na(scientificNameAuthorship)) %>%
@@ -40,7 +36,7 @@ preprocess_col <- function(url = paste0("http://www.catalogueoflife.org/DCA_Expo
              stri_trim(stri_replace_first_fixed(scientificName, scientificNameAuthorship, ""))),
     taxa_tmp %>%
       filter(is.na(scientificNameAuthorship))
-  )
+  ) 
 
   ## For accepted names, set acceptedNameUsageID to match taxonID, rather NA
   accepted <-
@@ -53,9 +49,9 @@ preprocess_col <- function(url = paste0("http://www.catalogueoflife.org/DCA_Expo
     filter(!is.na(acceptedNameUsageID)) %>%
     select(taxonID, scientificName, acceptedNameUsageID, taxonomicStatus) %>%
     left_join(accepted_heirarchy, by = c("acceptedNameUsageID" = "taxonID"))
+  
 
   # We drop un-mapped synonyms, as they are not helpful
-
 
   vernacular <- read_tsv(file.path(dir, "vernacular.txt"))
   #First we create the separate common names table
@@ -63,8 +59,8 @@ preprocess_col <- function(url = paste0("http://www.catalogueoflife.org/DCA_Expo
     select(taxonID, vernacularName, language) %>%
     inner_join(bind_rows(accepted), by = "taxonID") %>%
     mutate(taxonID = stringi::stri_paste("COL:", taxonID),
-           acceptedNameUsageID = stringi::stri_paste("COL:", acceptedNameUsageID))
-
+           acceptedNameUsageID = stringi::stri_paste("COL:", acceptedNameUsageID)
+           )
 
   # Also add a common name to the master dwc table
   # first english names,
@@ -79,21 +75,20 @@ preprocess_col <- function(url = paste0("http://www.catalogueoflife.org/DCA_Expo
[…]
