+++ data-raw/ott.R
@@ -1,9 +1,5 @@
-library(tidyverse)
-library(fs)
-source(here::here("data-raw/helper-routines.R"))
-
-
-
+#' @import fs tidyverse
+#' @export
 preprocess_ott <-
   function(url = "http://files.opentreeoflife.org/ott/ott3.2/ott3.2.tgz",
            output_paths = c(dwc = "2019/dwc_ott.tsv.bz2",
@@ -24,168 +20,171 @@ preprocess_ott <-
   dir <- fs::dir_ls(basedir, type="dir")
 
   ## defaulting to logicals is so annoying!
-  read_tsv <- function(...)
+  read_tsv <- function(...){
     readr::read_tsv(..., quote = "",
                     col_types = readr::cols(.default = "c"))
-
-
-# Really would be nice if we followed the DAMN STANDARD. tsv is tab-delimited,
-# not "\\t\\|\\t" delimited, people!
-
-## we `select()` in order to drop columns consisting solely of pipe separators...
-## ignore warnings
-synonyms <- read_tsv(file.path(dir, "synonyms.tsv")) %>%
-  select(name, uid, type, uniqname, sourceinfo)
-taxonomy <- read_tsv(file.path(dir, "taxonomy.tsv")) %>%
-  select(name, uid, parent_uid, rank, uniqname, sourceinfo, flags)
-
-
-## sourceinfo is comma-separated list of identifiers which synonym resolves against
-## (identifiers of accepted names, not just ids to the synonym, not listed)
-## UIDs are OTT ids of the ACCEPTED NAMES.  no ids to synonym names
-
-# synonyms involve a lot of types, but mostly "synonym".
-## Does not include "accepted" names.
-## synonyms %>% count(type) %>% arrange(desc(n))
-
-# taxonomy includes a lot of different flags,
-# including "extinct", "environmental", & "incertae_sedis"
-## taxonomy %>% count(flags) %>% arrange(desc(n))
-
-## DEPRECATED Synonyms table: id, accepted_name, rank, name, name_type
-#ott_synonyms <- taxonomy %>%
-#  select(accepted_name = name, uid, rank) %>%
-#  right_join(synonyms) %>%
-#  select(id = uid, accepted_name, name, rank, name_type = type) %>%
-#  mutate(id = paste0("OTT:", id))
-
-## TaxonID table
-ott_taxonid <- bind_rows(
-  taxonomy %>% select(id = uid, name, rank) %>%
-    mutate(id = paste0("OTT:", id)) %>%
-    mutate(accepted_id = id, name_type = "accepted_name"),
-  synonyms %>%
-    select(accepted_id = uid, name, name_type = type) %>%
-    left_join(select(taxonomy, uid, rank),
-              by = c("accepted_id" = "uid")) %>%
-    mutate(id = NA, accepted_id = paste0("OTT:", accepted_id))
-  )  %>% ## and rename to Darwin Core fields
-  rename(taxonID = id,
-         scientificName = name,
-         taxonRank = rank,
-         taxonomicStatus = name_type,
-         acceptedNameUsageID = accepted_id)
-
-
-rm(synonyms)
-
-## really slow but works and drops duplicates that distinct() does not
-ott_taxonid <- ott_taxonid %>%
-  de_duplicate()
-
-
-
-
-max <- pull(taxonomy, rank) %>% unique() %>% length()
-## Time to unpack another recursive taxonomy hierarchy
-ids <- select(taxonomy, id = uid, parent = parent_uid)
-hierarchy <- ids
-for(i in 1:max){
-  p <- paste0("p",i)
-  n <- names(hierarchy)
-  names(hierarchy) <- gsub("parent", p, n)
[…]
