### 2020-02-18 #344 @maelle
The badges server seems to have problems at the moment but yep this should be closed for the badge to become green I think.
